@@ -1,9 +1,9 @@
 module agones.dev/agones
 
-go 1.17
+go 1.19
 
 require (
-	cloud.google.com/go v0.54.0
+	cloud.google.com/go v0.81.0
 	contrib.go.opencensus.io/exporter/prometheus v0.1.0
 	contrib.go.opencensus.io/exporter/stackdriver v0.8.0
 	fortio.org/fortio v1.3.1
