@@ -1,5 +1,5 @@
 module github.com/agones/agones/site
 
-go 1.13
+go 1.19
 
 require gopkg.in/yaml.v2 v2.2.8
