@@ -21,7 +21,7 @@ import (
 	"crypto/x509"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"math"
 	"net/http"
 	"net/url"
@@ -35,8 +35,12 @@ import (
 	"k8s.io/apimachinery/pkg/util/uuid"
 )
 
+var tlsConfig = &tls.Config{}
 var client = http.Client{
 	Timeout: 15 * time.Second,
+	Transport: &http.Transport{
+		TLSClientConfig: tlsConfig,
+	},
 }
 
 // computeDesiredFleetSize computes the new desired size of the given fleet
@@ -115,11 +119,7 @@ func setCABundle(caBundle []byte) error {
 	if ok := rootCAs.AppendCertsFromPEM(caBundle); !ok {
 		return errors.New("no certs were appended from caBundle")
 	}
-	client.Transport = &http.Transport{
-		TLSClientConfig: &tls.Config{
-			RootCAs: rootCAs,
-		},
-	}
+	tlsConfig.RootCAs = rootCAs
 	return nil
 }
 
@@ -173,7 +173,7 @@ func applyWebhookPolicy(w *autoscalingv1.WebhookPolicy, f *agonesv1.Fleet) (repl
 	if res.StatusCode != http.StatusOK {
 		return 0, false, fmt.Errorf("bad status code %d from the server: %s", res.StatusCode, u.String())
 	}
-	result, err := ioutil.ReadAll(res.Body)
+	result, err := io.ReadAll(res.Body)
 	if err != nil {
 		return 0, false, err
 	}
