@@ -16,7 +16,7 @@ package gameserverallocations
 
 import (
 	"context"
-	"io/ioutil"
+	"io"
 	"mime"
 	"net/http"
 	"time"
@@ -172,7 +172,7 @@ func (c *Controller) allocationDeserialization(r *http.Request, namespace string
 		return gsa, errors.New("Could not find deserializer")
 	}
 
-	b, err := ioutil.ReadAll(r.Body)
+	b, err := io.ReadAll(r.Body)
 	if err != nil {
 		return gsa, errors.Wrap(err, "could not read body")
 	}
