@@ -18,8 +18,8 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"io/ioutil"
 	"net"
+	"os"
 	"os/exec"
 	"strings"
 	"sync/atomic"
@@ -399,11 +399,14 @@ func TestGameServerUnhealthyAfterReadyCrash(t *testing.T) {
 				l.Info("UDP Crash stop signal received. Stopping.")
 				return
 			}
-			conn, err := net.Dial("udp", address)
-			assert.NoError(t, err)
-			defer conn.Close() // nolint: errcheck
-			_, err = conn.Write([]byte("CRASH"))
-			if err != nil {
+			var writeErr error
+			func() {
+				conn, err := net.Dial("udp", address)
+				assert.NoError(t, err)
+				defer conn.Close() // nolint: errcheck
+				_, writeErr = conn.Write([]byte("CRASH"))
+			}()
+			if writeErr != nil {
 				l.WithError(err).Warn("error sending udp packet. Stopping.")
 				return
 			}
@@ -923,9 +926,9 @@ spec:
           preferredDuringSchedulingIgnoredDuringExecution: ERROR
       containers:
         - name: simple-game-server
-          image: gcr.io/agones-images/simple-game-server:0.13
+          image: gcr.io/agones-images/simple-game-server:0.14
 `
-	err := ioutil.WriteFile("/tmp/invalid.yaml", []byte(gsYaml), 0o644)
+	err := os.WriteFile("/tmp/invalid.yaml", []byte(gsYaml), 0o644)
 	require.NoError(t, err)
 
 	cmd := exec.Command("kubectl", "apply", "-f", "/tmp/invalid.yaml")
