@@ -118,7 +118,7 @@ func (c *FakeGameServerSets) UpdateStatus(ctx context.Context, gameServerSet *ag
 // Delete takes name of the gameServerSet and deletes it. Returns an error if one occurs.
 func (c *FakeGameServerSets) Delete(ctx context.Context, name string, opts v1.DeleteOptions) error {
 	_, err := c.Fake.
-		Invokes(testing.NewDeleteAction(gameserversetsResource, c.ns, name), &agonesv1.GameServerSet{})
+		Invokes(testing.NewDeleteActionWithOptions(gameserversetsResource, c.ns, name, opts), &agonesv1.GameServerSet{})
 
 	return err
 }
