@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC All Rights Reserved.
+// Copyright 2022 Google LLC All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,12 +23,12 @@
 #include <string>
 
 #include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3011000
+#if PROTOBUF_VERSION < 3014000
 #error This file was generated by a newer version of protoc which is
 #error incompatible with your Protocol Buffer headers. Please update
 #error your headers.
 #endif
-#if 3011002 < PROTOBUF_MIN_PROTOC_VERSION
+#if 3014000 < PROTOBUF_MIN_PROTOC_VERSION
 #error This file was generated by an older version of protoc which is
 #error incompatible with your Protocol Buffer headers. Please
 #error regenerate this file with a newer version of protoc.
@@ -40,8 +40,7 @@
 #include <google/protobuf/arenastring.h>
 #include <google/protobuf/generated_message_table_driven.h>
 #include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata.h>
+#include <google/protobuf/metadata_lite.h>
 #include <google/protobuf/generated_message_reflection.h>
 #include <google/protobuf/message.h>
 #include <google/protobuf/repeated_field.h>  // IWYU pragma: export
@@ -64,7 +63,7 @@ PROTOBUF_NAMESPACE_CLOSE
 struct AGONES_EXPORT TableStruct_sdk_2eproto {
   static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
     PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
+  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxiliaryParseTableField aux[]
     PROTOBUF_SECTION_VARIABLE(protodesc_cold);
   static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[12]
     PROTOBUF_SECTION_VARIABLE(protodesc_cold);
@@ -135,10 +134,10 @@ namespace sdk {
 
 // ===================================================================
 
-class AGONES_EXPORT Empty :
+class AGONES_EXPORT Empty PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.Empty) */ {
  public:
-  Empty();
+  inline Empty() : Empty(nullptr) {}
   virtual ~Empty();
 
   Empty(const Empty& from);
@@ -152,7 +151,7 @@ class AGONES_EXPORT Empty :
     return *this;
   }
   inline Empty& operator=(Empty&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -171,7 +170,6 @@ class AGONES_EXPORT Empty :
   }
   static const Empty& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const Empty* internal_default_instance() {
     return reinterpret_cast<const Empty*>(
                &_Empty_default_instance_);
@@ -184,6 +182,15 @@ class AGONES_EXPORT Empty :
   }
   inline void Swap(Empty* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(Empty* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -218,13 +225,11 @@ class AGONES_EXPORT Empty :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.Empty";
   }
+  protected:
+  explicit Empty(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -244,16 +249,18 @@ class AGONES_EXPORT Empty :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
   friend struct ::TableStruct_sdk_2eproto;
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT KeyValue :
+class AGONES_EXPORT KeyValue PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.KeyValue) */ {
  public:
-  KeyValue();
+  inline KeyValue() : KeyValue(nullptr) {}
   virtual ~KeyValue();
 
   KeyValue(const KeyValue& from);
@@ -267,7 +274,7 @@ class AGONES_EXPORT KeyValue :
     return *this;
   }
   inline KeyValue& operator=(KeyValue&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -286,7 +293,6 @@ class AGONES_EXPORT KeyValue :
   }
   static const KeyValue& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const KeyValue* internal_default_instance() {
     return reinterpret_cast<const KeyValue*>(
                &_KeyValue_default_instance_);
@@ -299,6 +305,15 @@ class AGONES_EXPORT KeyValue :
   }
   inline void Swap(KeyValue* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(KeyValue* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -333,13 +348,11 @@ class AGONES_EXPORT KeyValue :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.KeyValue";
   }
+  protected:
+  explicit KeyValue(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -395,7 +408,9 @@ class AGONES_EXPORT KeyValue :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr key_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr value_;
   mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
@@ -403,10 +418,10 @@ class AGONES_EXPORT KeyValue :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT Duration :
+class AGONES_EXPORT Duration PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.Duration) */ {
  public:
-  Duration();
+  inline Duration() : Duration(nullptr) {}
   virtual ~Duration();
 
   Duration(const Duration& from);
@@ -420,7 +435,7 @@ class AGONES_EXPORT Duration :
     return *this;
   }
   inline Duration& operator=(Duration&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -439,7 +454,6 @@ class AGONES_EXPORT Duration :
   }
   static const Duration& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const Duration* internal_default_instance() {
     return reinterpret_cast<const Duration*>(
                &_Duration_default_instance_);
@@ -452,6 +466,15 @@ class AGONES_EXPORT Duration :
   }
   inline void Swap(Duration* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(Duration* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -486,13 +509,11 @@ class AGONES_EXPORT Duration :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.Duration";
   }
+  protected:
+  explicit Duration(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -524,7 +545,9 @@ class AGONES_EXPORT Duration :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::PROTOBUF_NAMESPACE_ID::int64 seconds_;
   mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
   friend struct ::TableStruct_sdk_2eproto;
@@ -534,16 +557,14 @@ class AGONES_EXPORT Duration :
 class GameServer_ObjectMeta_AnnotationsEntry_DoNotUse : public ::PROTOBUF_NAMESPACE_ID::internal::MapEntry<GameServer_ObjectMeta_AnnotationsEntry_DoNotUse, 
     std::string, std::string,
     ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    0 > {
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING> {
 public:
   typedef ::PROTOBUF_NAMESPACE_ID::internal::MapEntry<GameServer_ObjectMeta_AnnotationsEntry_DoNotUse, 
     std::string, std::string,
     ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    0 > SuperType;
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING> SuperType;
   GameServer_ObjectMeta_AnnotationsEntry_DoNotUse();
-  GameServer_ObjectMeta_AnnotationsEntry_DoNotUse(::PROTOBUF_NAMESPACE_ID::Arena* arena);
+  explicit GameServer_ObjectMeta_AnnotationsEntry_DoNotUse(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   void MergeFrom(const GameServer_ObjectMeta_AnnotationsEntry_DoNotUse& other);
   static const GameServer_ObjectMeta_AnnotationsEntry_DoNotUse* internal_default_instance() { return reinterpret_cast<const GameServer_ObjectMeta_AnnotationsEntry_DoNotUse*>(&_GameServer_ObjectMeta_AnnotationsEntry_DoNotUse_default_instance_); }
   static bool ValidateKey(std::string* s) {
@@ -568,16 +589,14 @@ public:
 class GameServer_ObjectMeta_LabelsEntry_DoNotUse : public ::PROTOBUF_NAMESPACE_ID::internal::MapEntry<GameServer_ObjectMeta_LabelsEntry_DoNotUse, 
     std::string, std::string,
     ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    0 > {
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING> {
 public:
   typedef ::PROTOBUF_NAMESPACE_ID::internal::MapEntry<GameServer_ObjectMeta_LabelsEntry_DoNotUse, 
     std::string, std::string,
     ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-    0 > SuperType;
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING> SuperType;
   GameServer_ObjectMeta_LabelsEntry_DoNotUse();
-  GameServer_ObjectMeta_LabelsEntry_DoNotUse(::PROTOBUF_NAMESPACE_ID::Arena* arena);
+  explicit GameServer_ObjectMeta_LabelsEntry_DoNotUse(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   void MergeFrom(const GameServer_ObjectMeta_LabelsEntry_DoNotUse& other);
   static const GameServer_ObjectMeta_LabelsEntry_DoNotUse* internal_default_instance() { return reinterpret_cast<const GameServer_ObjectMeta_LabelsEntry_DoNotUse*>(&_GameServer_ObjectMeta_LabelsEntry_DoNotUse_default_instance_); }
   static bool ValidateKey(std::string* s) {
@@ -599,10 +618,10 @@ public:
 
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer_ObjectMeta :
+class AGONES_EXPORT GameServer_ObjectMeta PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer.ObjectMeta) */ {
  public:
-  GameServer_ObjectMeta();
+  inline GameServer_ObjectMeta() : GameServer_ObjectMeta(nullptr) {}
   virtual ~GameServer_ObjectMeta();
 
   GameServer_ObjectMeta(const GameServer_ObjectMeta& from);
@@ -616,7 +635,7 @@ class AGONES_EXPORT GameServer_ObjectMeta :
     return *this;
   }
   inline GameServer_ObjectMeta& operator=(GameServer_ObjectMeta&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -635,7 +654,6 @@ class AGONES_EXPORT GameServer_ObjectMeta :
   }
   static const GameServer_ObjectMeta& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer_ObjectMeta* internal_default_instance() {
     return reinterpret_cast<const GameServer_ObjectMeta*>(
                &_GameServer_ObjectMeta_default_instance_);
@@ -648,6 +666,15 @@ class AGONES_EXPORT GameServer_ObjectMeta :
   }
   inline void Swap(GameServer_ObjectMeta* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer_ObjectMeta* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -682,13 +709,11 @@ class AGONES_EXPORT GameServer_ObjectMeta :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer.ObjectMeta";
   }
+  protected:
+  explicit GameServer_ObjectMeta(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -845,19 +870,19 @@ class AGONES_EXPORT GameServer_ObjectMeta :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::PROTOBUF_NAMESPACE_ID::internal::MapField<
       GameServer_ObjectMeta_AnnotationsEntry_DoNotUse,
       std::string, std::string,
       ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-      0 > annotations_;
+      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING> annotations_;
   ::PROTOBUF_NAMESPACE_ID::internal::MapField<
       GameServer_ObjectMeta_LabelsEntry_DoNotUse,
       std::string, std::string,
       ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING,
-      0 > labels_;
+      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_STRING> labels_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr name_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr namespace__;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr uid_;
@@ -870,10 +895,10 @@ class AGONES_EXPORT GameServer_ObjectMeta :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer_Spec_Health :
+class AGONES_EXPORT GameServer_Spec_Health PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer.Spec.Health) */ {
  public:
-  GameServer_Spec_Health();
+  inline GameServer_Spec_Health() : GameServer_Spec_Health(nullptr) {}
   virtual ~GameServer_Spec_Health();
 
   GameServer_Spec_Health(const GameServer_Spec_Health& from);
@@ -887,7 +912,7 @@ class AGONES_EXPORT GameServer_Spec_Health :
     return *this;
   }
   inline GameServer_Spec_Health& operator=(GameServer_Spec_Health&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -906,7 +931,6 @@ class AGONES_EXPORT GameServer_Spec_Health :
   }
   static const GameServer_Spec_Health& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer_Spec_Health* internal_default_instance() {
     return reinterpret_cast<const GameServer_Spec_Health*>(
                &_GameServer_Spec_Health_default_instance_);
@@ -919,6 +943,15 @@ class AGONES_EXPORT GameServer_Spec_Health :
   }
   inline void Swap(GameServer_Spec_Health* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer_Spec_Health* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -953,13 +986,11 @@ class AGONES_EXPORT GameServer_Spec_Health :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer.Spec.Health";
   }
+  protected:
+  explicit GameServer_Spec_Health(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -1021,7 +1052,9 @@ class AGONES_EXPORT GameServer_Spec_Health :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   bool disabled_;
   ::PROTOBUF_NAMESPACE_ID::int32 period_seconds_;
   ::PROTOBUF_NAMESPACE_ID::int32 failure_threshold_;
@@ -1031,10 +1064,10 @@ class AGONES_EXPORT GameServer_Spec_Health :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer_Spec :
+class AGONES_EXPORT GameServer_Spec PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer.Spec) */ {
  public:
-  GameServer_Spec();
+  inline GameServer_Spec() : GameServer_Spec(nullptr) {}
   virtual ~GameServer_Spec();
 
   GameServer_Spec(const GameServer_Spec& from);
@@ -1048,7 +1081,7 @@ class AGONES_EXPORT GameServer_Spec :
     return *this;
   }
   inline GameServer_Spec& operator=(GameServer_Spec&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -1067,7 +1100,6 @@ class AGONES_EXPORT GameServer_Spec :
   }
   static const GameServer_Spec& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer_Spec* internal_default_instance() {
     return reinterpret_cast<const GameServer_Spec*>(
                &_GameServer_Spec_default_instance_);
@@ -1080,6 +1112,15 @@ class AGONES_EXPORT GameServer_Spec :
   }
   inline void Swap(GameServer_Spec* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer_Spec* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -1114,13 +1155,11 @@ class AGONES_EXPORT GameServer_Spec :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer.Spec";
   }
+  protected:
+  explicit GameServer_Spec(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -1155,22 +1194,27 @@ class AGONES_EXPORT GameServer_Spec :
   const ::agones::dev::sdk::GameServer_Spec_Health& _internal_health() const;
   ::agones::dev::sdk::GameServer_Spec_Health* _internal_mutable_health();
   public:
+  void unsafe_arena_set_allocated_health(
+      ::agones::dev::sdk::GameServer_Spec_Health* health);
+  ::agones::dev::sdk::GameServer_Spec_Health* unsafe_arena_release_health();
 
   // @@protoc_insertion_point(class_scope:agones.dev.sdk.GameServer.Spec)
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::agones::dev::sdk::GameServer_Spec_Health* health_;
   mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
   friend struct ::TableStruct_sdk_2eproto;
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer_Status_Port :
+class AGONES_EXPORT GameServer_Status_Port PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer.Status.Port) */ {
  public:
-  GameServer_Status_Port();
+  inline GameServer_Status_Port() : GameServer_Status_Port(nullptr) {}
   virtual ~GameServer_Status_Port();
 
   GameServer_Status_Port(const GameServer_Status_Port& from);
@@ -1184,7 +1228,7 @@ class AGONES_EXPORT GameServer_Status_Port :
     return *this;
   }
   inline GameServer_Status_Port& operator=(GameServer_Status_Port&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -1203,7 +1247,6 @@ class AGONES_EXPORT GameServer_Status_Port :
   }
   static const GameServer_Status_Port& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer_Status_Port* internal_default_instance() {
     return reinterpret_cast<const GameServer_Status_Port*>(
                &_GameServer_Status_Port_default_instance_);
@@ -1216,6 +1259,15 @@ class AGONES_EXPORT GameServer_Status_Port :
   }
   inline void Swap(GameServer_Status_Port* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer_Status_Port* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -1250,13 +1302,11 @@ class AGONES_EXPORT GameServer_Status_Port :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer.Status.Port";
   }
+  protected:
+  explicit GameServer_Status_Port(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -1305,7 +1355,9 @@ class AGONES_EXPORT GameServer_Status_Port :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr name_;
   ::PROTOBUF_NAMESPACE_ID::int32 port_;
   mutable ::PROTOBUF_NAMESPACE_ID::internal::CachedSize _cached_size_;
@@ -1313,10 +1365,10 @@ class AGONES_EXPORT GameServer_Status_Port :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer_Status_PlayerStatus :
+class AGONES_EXPORT GameServer_Status_PlayerStatus PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer.Status.PlayerStatus) */ {
  public:
-  GameServer_Status_PlayerStatus();
+  inline GameServer_Status_PlayerStatus() : GameServer_Status_PlayerStatus(nullptr) {}
   virtual ~GameServer_Status_PlayerStatus();
 
   GameServer_Status_PlayerStatus(const GameServer_Status_PlayerStatus& from);
@@ -1330,7 +1382,7 @@ class AGONES_EXPORT GameServer_Status_PlayerStatus :
     return *this;
   }
   inline GameServer_Status_PlayerStatus& operator=(GameServer_Status_PlayerStatus&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -1349,7 +1401,6 @@ class AGONES_EXPORT GameServer_Status_PlayerStatus :
   }
   static const GameServer_Status_PlayerStatus& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer_Status_PlayerStatus* internal_default_instance() {
     return reinterpret_cast<const GameServer_Status_PlayerStatus*>(
                &_GameServer_Status_PlayerStatus_default_instance_);
@@ -1362,6 +1413,15 @@ class AGONES_EXPORT GameServer_Status_PlayerStatus :
   }
   inline void Swap(GameServer_Status_PlayerStatus* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer_Status_PlayerStatus* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -1396,13 +1456,11 @@ class AGONES_EXPORT GameServer_Status_PlayerStatus :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer.Status.PlayerStatus";
   }
+  protected:
+  explicit GameServer_Status_PlayerStatus(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -1469,7 +1527,9 @@ class AGONES_EXPORT GameServer_Status_PlayerStatus :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::PROTOBUF_NAMESPACE_ID::RepeatedPtrField<std::string> ids_;
   ::PROTOBUF_NAMESPACE_ID::int64 count_;
   ::PROTOBUF_NAMESPACE_ID::int64 capacity_;
@@ -1478,10 +1538,10 @@ class AGONES_EXPORT GameServer_Status_PlayerStatus :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer_Status :
+class AGONES_EXPORT GameServer_Status PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer.Status) */ {
  public:
-  GameServer_Status();
+  inline GameServer_Status() : GameServer_Status(nullptr) {}
   virtual ~GameServer_Status();
 
   GameServer_Status(const GameServer_Status& from);
@@ -1495,7 +1555,7 @@ class AGONES_EXPORT GameServer_Status :
     return *this;
   }
   inline GameServer_Status& operator=(GameServer_Status&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -1514,7 +1574,6 @@ class AGONES_EXPORT GameServer_Status :
   }
   static const GameServer_Status& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer_Status* internal_default_instance() {
     return reinterpret_cast<const GameServer_Status*>(
                &_GameServer_Status_default_instance_);
@@ -1527,6 +1586,15 @@ class AGONES_EXPORT GameServer_Status :
   }
   inline void Swap(GameServer_Status* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer_Status* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -1561,13 +1629,11 @@ class AGONES_EXPORT GameServer_Status :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer.Status";
   }
+  protected:
+  explicit GameServer_Status(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -1656,12 +1722,17 @@ class AGONES_EXPORT GameServer_Status :
   const ::agones::dev::sdk::GameServer_Status_PlayerStatus& _internal_players() const;
   ::agones::dev::sdk::GameServer_Status_PlayerStatus* _internal_mutable_players();
   public:
+  void unsafe_arena_set_allocated_players(
+      ::agones::dev::sdk::GameServer_Status_PlayerStatus* players);
+  ::agones::dev::sdk::GameServer_Status_PlayerStatus* unsafe_arena_release_players();
 
   // @@protoc_insertion_point(class_scope:agones.dev.sdk.GameServer.Status)
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::PROTOBUF_NAMESPACE_ID::RepeatedPtrField< ::agones::dev::sdk::GameServer_Status_Port > ports_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr state_;
   ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr address_;
@@ -1671,10 +1742,10 @@ class AGONES_EXPORT GameServer_Status :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT GameServer :
+class AGONES_EXPORT GameServer PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:agones.dev.sdk.GameServer) */ {
  public:
-  GameServer();
+  inline GameServer() : GameServer(nullptr) {}
   virtual ~GameServer();
 
   GameServer(const GameServer& from);
@@ -1688,7 +1759,7 @@ class AGONES_EXPORT GameServer :
     return *this;
   }
   inline GameServer& operator=(GameServer&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -1707,7 +1778,6 @@ class AGONES_EXPORT GameServer :
   }
   static const GameServer& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const GameServer* internal_default_instance() {
     return reinterpret_cast<const GameServer*>(
                &_GameServer_default_instance_);
@@ -1720,6 +1790,15 @@ class AGONES_EXPORT GameServer :
   }
   inline void Swap(GameServer* other) {
     if (other == this) return;
+    if (GetArena() == other->GetArena()) {
+      InternalSwap(other);
+    } else {
+      ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
+    }
+  }
+  void UnsafeArenaSwap(GameServer* other) {
+    if (other == this) return;
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -1754,13 +1833,11 @@ class AGONES_EXPORT GameServer :
   static ::PROTOBUF_NAMESPACE_ID::StringPiece FullMessageName() {
     return "agones.dev.sdk.GameServer";
   }
+  protected:
+  explicit GameServer(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return nullptr;
-  }
-  inline void* MaybeArenaPtr() const {
-    return nullptr;
-  }
+  static void ArenaDtor(void* object);
+  inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -1799,6 +1876,9 @@ class AGONES_EXPORT GameServer :
   const ::agones::dev::sdk::GameServer_ObjectMeta& _internal_object_meta() const;
   ::agones::dev::sdk::GameServer_ObjectMeta* _internal_mutable_object_meta();
   public:
+  void unsafe_arena_set_allocated_object_meta(
+      ::agones::dev::sdk::GameServer_ObjectMeta* object_meta);
+  ::agones::dev::sdk::GameServer_ObjectMeta* unsafe_arena_release_object_meta();
 
   // .agones.dev.sdk.GameServer.Spec spec = 2;
   bool has_spec() const;
@@ -1814,6 +1894,9 @@ class AGONES_EXPORT GameServer :
   const ::agones::dev::sdk::GameServer_Spec& _internal_spec() const;
   ::agones::dev::sdk::GameServer_Spec* _internal_mutable_spec();
   public:
+  void unsafe_arena_set_allocated_spec(
+      ::agones::dev::sdk::GameServer_Spec* spec);
+  ::agones::dev::sdk::GameServer_Spec* unsafe_arena_release_spec();
 
   // .agones.dev.sdk.GameServer.Status status = 3;
   bool has_status() const;
@@ -1829,12 +1912,17 @@ class AGONES_EXPORT GameServer :
   const ::agones::dev::sdk::GameServer_Status& _internal_status() const;
   ::agones::dev::sdk::GameServer_Status* _internal_mutable_status();
   public:
+  void unsafe_arena_set_allocated_status(
+      ::agones::dev::sdk::GameServer_Status* status);
+  ::agones::dev::sdk::GameServer_Status* unsafe_arena_release_status();
 
   // @@protoc_insertion_point(class_scope:agones.dev.sdk.GameServer)
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
+  template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
+  typedef void InternalArenaConstructable_;
+  typedef void DestructorSkippable_;
   ::agones::dev::sdk::GameServer_ObjectMeta* object_meta_;
   ::agones::dev::sdk::GameServer_Spec* spec_;
   ::agones::dev::sdk::GameServer_Status* status_;
@@ -1858,7 +1946,7 @@ class AGONES_EXPORT GameServer :
 
 // string key = 1;
 inline void KeyValue::clear_key() {
-  key_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  key_.ClearToEmpty();
 }
 inline const std::string& KeyValue::key() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.KeyValue.key)
@@ -1873,38 +1961,38 @@ inline std::string* KeyValue::mutable_key() {
   return _internal_mutable_key();
 }
 inline const std::string& KeyValue::_internal_key() const {
-  return key_.GetNoArena();
+  return key_.Get();
 }
 inline void KeyValue::_internal_set_key(const std::string& value) {
   
-  key_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  key_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void KeyValue::set_key(std::string&& value) {
   
-  key_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  key_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.KeyValue.key)
 }
 inline void KeyValue::set_key(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  key_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  key_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.KeyValue.key)
 }
-inline void KeyValue::set_key(const char* value, size_t size) {
+inline void KeyValue::set_key(const char* value,
+    size_t size) {
   
-  key_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  key_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.KeyValue.key)
 }
 inline std::string* KeyValue::_internal_mutable_key() {
   
-  return key_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return key_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* KeyValue::release_key() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.KeyValue.key)
-  
-  return key_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return key_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void KeyValue::set_allocated_key(std::string* key) {
   if (key != nullptr) {
@@ -1912,13 +2000,14 @@ inline void KeyValue::set_allocated_key(std::string* key) {
   } else {
     
   }
-  key_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), key);
+  key_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), key,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.KeyValue.key)
 }
 
 // string value = 2;
 inline void KeyValue::clear_value() {
-  value_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  value_.ClearToEmpty();
 }
 inline const std::string& KeyValue::value() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.KeyValue.value)
@@ -1933,38 +2022,38 @@ inline std::string* KeyValue::mutable_value() {
   return _internal_mutable_value();
 }
 inline const std::string& KeyValue::_internal_value() const {
-  return value_.GetNoArena();
+  return value_.Get();
 }
 inline void KeyValue::_internal_set_value(const std::string& value) {
   
-  value_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  value_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void KeyValue::set_value(std::string&& value) {
   
-  value_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  value_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.KeyValue.value)
 }
 inline void KeyValue::set_value(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  value_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  value_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.KeyValue.value)
 }
-inline void KeyValue::set_value(const char* value, size_t size) {
+inline void KeyValue::set_value(const char* value,
+    size_t size) {
   
-  value_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  value_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.KeyValue.value)
 }
 inline std::string* KeyValue::_internal_mutable_value() {
   
-  return value_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return value_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* KeyValue::release_value() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.KeyValue.value)
-  
-  return value_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return value_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void KeyValue::set_allocated_value(std::string* value) {
   if (value != nullptr) {
@@ -1972,7 +2061,8 @@ inline void KeyValue::set_allocated_value(std::string* value) {
   } else {
     
   }
-  value_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  value_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.KeyValue.value)
 }
 
@@ -2010,7 +2100,7 @@ inline void Duration::set_seconds(::PROTOBUF_NAMESPACE_ID::int64 value) {
 
 // string name = 1;
 inline void GameServer_ObjectMeta::clear_name() {
-  name_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  name_.ClearToEmpty();
 }
 inline const std::string& GameServer_ObjectMeta::name() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.ObjectMeta.name)
@@ -2025,38 +2115,38 @@ inline std::string* GameServer_ObjectMeta::mutable_name() {
   return _internal_mutable_name();
 }
 inline const std::string& GameServer_ObjectMeta::_internal_name() const {
-  return name_.GetNoArena();
+  return name_.Get();
 }
 inline void GameServer_ObjectMeta::_internal_set_name(const std::string& value) {
   
-  name_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  name_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_ObjectMeta::set_name(std::string&& value) {
   
-  name_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  name_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.ObjectMeta.name)
 }
 inline void GameServer_ObjectMeta::set_name(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  name_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  name_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.ObjectMeta.name)
 }
-inline void GameServer_ObjectMeta::set_name(const char* value, size_t size) {
+inline void GameServer_ObjectMeta::set_name(const char* value,
+    size_t size) {
   
-  name_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  name_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.ObjectMeta.name)
 }
 inline std::string* GameServer_ObjectMeta::_internal_mutable_name() {
   
-  return name_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return name_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_ObjectMeta::release_name() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.ObjectMeta.name)
-  
-  return name_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return name_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_ObjectMeta::set_allocated_name(std::string* name) {
   if (name != nullptr) {
@@ -2064,13 +2154,14 @@ inline void GameServer_ObjectMeta::set_allocated_name(std::string* name) {
   } else {
     
   }
-  name_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), name);
+  name_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), name,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.ObjectMeta.name)
 }
 
 // string namespace = 2;
 inline void GameServer_ObjectMeta::clear_namespace_() {
-  namespace__.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  namespace__.ClearToEmpty();
 }
 inline const std::string& GameServer_ObjectMeta::namespace_() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.ObjectMeta.namespace)
@@ -2085,38 +2176,38 @@ inline std::string* GameServer_ObjectMeta::mutable_namespace_() {
   return _internal_mutable_namespace_();
 }
 inline const std::string& GameServer_ObjectMeta::_internal_namespace_() const {
-  return namespace__.GetNoArena();
+  return namespace__.Get();
 }
 inline void GameServer_ObjectMeta::_internal_set_namespace_(const std::string& value) {
   
-  namespace__.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  namespace__.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_ObjectMeta::set_namespace_(std::string&& value) {
   
-  namespace__.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  namespace__.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.ObjectMeta.namespace)
 }
 inline void GameServer_ObjectMeta::set_namespace_(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  namespace__.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  namespace__.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.ObjectMeta.namespace)
 }
-inline void GameServer_ObjectMeta::set_namespace_(const char* value, size_t size) {
+inline void GameServer_ObjectMeta::set_namespace_(const char* value,
+    size_t size) {
   
-  namespace__.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  namespace__.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.ObjectMeta.namespace)
 }
 inline std::string* GameServer_ObjectMeta::_internal_mutable_namespace_() {
   
-  return namespace__.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return namespace__.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_ObjectMeta::release_namespace_() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.ObjectMeta.namespace)
-  
-  return namespace__.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return namespace__.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_ObjectMeta::set_allocated_namespace_(std::string* namespace_) {
   if (namespace_ != nullptr) {
@@ -2124,13 +2215,14 @@ inline void GameServer_ObjectMeta::set_allocated_namespace_(std::string* namespa
   } else {
     
   }
-  namespace__.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), namespace_);
+  namespace__.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), namespace_,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.ObjectMeta.namespace)
 }
 
 // string uid = 3;
 inline void GameServer_ObjectMeta::clear_uid() {
-  uid_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  uid_.ClearToEmpty();
 }
 inline const std::string& GameServer_ObjectMeta::uid() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.ObjectMeta.uid)
@@ -2145,38 +2237,38 @@ inline std::string* GameServer_ObjectMeta::mutable_uid() {
   return _internal_mutable_uid();
 }
 inline const std::string& GameServer_ObjectMeta::_internal_uid() const {
-  return uid_.GetNoArena();
+  return uid_.Get();
 }
 inline void GameServer_ObjectMeta::_internal_set_uid(const std::string& value) {
   
-  uid_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  uid_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_ObjectMeta::set_uid(std::string&& value) {
   
-  uid_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  uid_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.ObjectMeta.uid)
 }
 inline void GameServer_ObjectMeta::set_uid(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  uid_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  uid_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.ObjectMeta.uid)
 }
-inline void GameServer_ObjectMeta::set_uid(const char* value, size_t size) {
+inline void GameServer_ObjectMeta::set_uid(const char* value,
+    size_t size) {
   
-  uid_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  uid_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.ObjectMeta.uid)
 }
 inline std::string* GameServer_ObjectMeta::_internal_mutable_uid() {
   
-  return uid_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return uid_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_ObjectMeta::release_uid() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.ObjectMeta.uid)
-  
-  return uid_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return uid_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_ObjectMeta::set_allocated_uid(std::string* uid) {
   if (uid != nullptr) {
@@ -2184,13 +2276,14 @@ inline void GameServer_ObjectMeta::set_allocated_uid(std::string* uid) {
   } else {
     
   }
-  uid_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), uid);
+  uid_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), uid,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.ObjectMeta.uid)
 }
 
 // string resource_version = 4;
 inline void GameServer_ObjectMeta::clear_resource_version() {
-  resource_version_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  resource_version_.ClearToEmpty();
 }
 inline const std::string& GameServer_ObjectMeta::resource_version() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.ObjectMeta.resource_version)
@@ -2205,38 +2298,38 @@ inline std::string* GameServer_ObjectMeta::mutable_resource_version() {
   return _internal_mutable_resource_version();
 }
 inline const std::string& GameServer_ObjectMeta::_internal_resource_version() const {
-  return resource_version_.GetNoArena();
+  return resource_version_.Get();
 }
 inline void GameServer_ObjectMeta::_internal_set_resource_version(const std::string& value) {
   
-  resource_version_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  resource_version_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_ObjectMeta::set_resource_version(std::string&& value) {
   
-  resource_version_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  resource_version_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.ObjectMeta.resource_version)
 }
 inline void GameServer_ObjectMeta::set_resource_version(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  resource_version_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  resource_version_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.ObjectMeta.resource_version)
 }
-inline void GameServer_ObjectMeta::set_resource_version(const char* value, size_t size) {
+inline void GameServer_ObjectMeta::set_resource_version(const char* value,
+    size_t size) {
   
-  resource_version_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  resource_version_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.ObjectMeta.resource_version)
 }
 inline std::string* GameServer_ObjectMeta::_internal_mutable_resource_version() {
   
-  return resource_version_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return resource_version_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_ObjectMeta::release_resource_version() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.ObjectMeta.resource_version)
-  
-  return resource_version_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return resource_version_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_ObjectMeta::set_allocated_resource_version(std::string* resource_version) {
   if (resource_version != nullptr) {
@@ -2244,7 +2337,8 @@ inline void GameServer_ObjectMeta::set_allocated_resource_version(std::string* r
   } else {
     
   }
-  resource_version_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), resource_version);
+  resource_version_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), resource_version,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.ObjectMeta.resource_version)
 }
 
@@ -2462,21 +2556,43 @@ inline bool GameServer_Spec::has_health() const {
   return _internal_has_health();
 }
 inline void GameServer_Spec::clear_health() {
-  if (GetArenaNoVirtual() == nullptr && health_ != nullptr) {
+  if (GetArena() == nullptr && health_ != nullptr) {
     delete health_;
   }
   health_ = nullptr;
 }
 inline const ::agones::dev::sdk::GameServer_Spec_Health& GameServer_Spec::_internal_health() const {
   const ::agones::dev::sdk::GameServer_Spec_Health* p = health_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::agones::dev::sdk::GameServer_Spec_Health*>(
-      &::agones::dev::sdk::_GameServer_Spec_Health_default_instance_);
+  return p != nullptr ? *p : reinterpret_cast<const ::agones::dev::sdk::GameServer_Spec_Health&>(
+      ::agones::dev::sdk::_GameServer_Spec_Health_default_instance_);
 }
 inline const ::agones::dev::sdk::GameServer_Spec_Health& GameServer_Spec::health() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.Spec.health)
   return _internal_health();
 }
+inline void GameServer_Spec::unsafe_arena_set_allocated_health(
+    ::agones::dev::sdk::GameServer_Spec_Health* health) {
+  if (GetArena() == nullptr) {
+    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(health_);
+  }
+  health_ = health;
+  if (health) {
+    
+  } else {
+    
+  }
+  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:agones.dev.sdk.GameServer.Spec.health)
+}
 inline ::agones::dev::sdk::GameServer_Spec_Health* GameServer_Spec::release_health() {
+  
+  ::agones::dev::sdk::GameServer_Spec_Health* temp = health_;
+  health_ = nullptr;
+  if (GetArena() != nullptr) {
+    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  }
+  return temp;
+}
+inline ::agones::dev::sdk::GameServer_Spec_Health* GameServer_Spec::unsafe_arena_release_health() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.Spec.health)
   
   ::agones::dev::sdk::GameServer_Spec_Health* temp = health_;
@@ -2486,7 +2602,7 @@ inline ::agones::dev::sdk::GameServer_Spec_Health* GameServer_Spec::release_heal
 inline ::agones::dev::sdk::GameServer_Spec_Health* GameServer_Spec::_internal_mutable_health() {
   
   if (health_ == nullptr) {
-    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Spec_Health>(GetArenaNoVirtual());
+    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Spec_Health>(GetArena());
     health_ = p;
   }
   return health_;
@@ -2496,12 +2612,13 @@ inline ::agones::dev::sdk::GameServer_Spec_Health* GameServer_Spec::mutable_heal
   return _internal_mutable_health();
 }
 inline void GameServer_Spec::set_allocated_health(::agones::dev::sdk::GameServer_Spec_Health* health) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArenaNoVirtual();
+  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
   if (message_arena == nullptr) {
     delete health_;
   }
   if (health) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena = nullptr;
+    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
+      ::PROTOBUF_NAMESPACE_ID::Arena::GetArena(health);
     if (message_arena != submessage_arena) {
       health = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
           message_arena, health, submessage_arena);
@@ -2520,7 +2637,7 @@ inline void GameServer_Spec::set_allocated_health(::agones::dev::sdk::GameServer
 
 // string name = 1;
 inline void GameServer_Status_Port::clear_name() {
-  name_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  name_.ClearToEmpty();
 }
 inline const std::string& GameServer_Status_Port::name() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.Status.Port.name)
@@ -2535,38 +2652,38 @@ inline std::string* GameServer_Status_Port::mutable_name() {
   return _internal_mutable_name();
 }
 inline const std::string& GameServer_Status_Port::_internal_name() const {
-  return name_.GetNoArena();
+  return name_.Get();
 }
 inline void GameServer_Status_Port::_internal_set_name(const std::string& value) {
   
-  name_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  name_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_Status_Port::set_name(std::string&& value) {
   
-  name_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  name_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.Status.Port.name)
 }
 inline void GameServer_Status_Port::set_name(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  name_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  name_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.Status.Port.name)
 }
-inline void GameServer_Status_Port::set_name(const char* value, size_t size) {
+inline void GameServer_Status_Port::set_name(const char* value,
+    size_t size) {
   
-  name_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  name_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.Status.Port.name)
 }
 inline std::string* GameServer_Status_Port::_internal_mutable_name() {
   
-  return name_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return name_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_Status_Port::release_name() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.Status.Port.name)
-  
-  return name_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return name_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_Status_Port::set_allocated_name(std::string* name) {
   if (name != nullptr) {
@@ -2574,7 +2691,8 @@ inline void GameServer_Status_Port::set_allocated_name(std::string* name) {
   } else {
     
   }
-  name_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), name);
+  name_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), name,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.Status.Port.name)
 }
 
@@ -2722,7 +2840,7 @@ GameServer_Status_PlayerStatus::mutable_ids() {
 
 // string state = 1;
 inline void GameServer_Status::clear_state() {
-  state_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  state_.ClearToEmpty();
 }
 inline const std::string& GameServer_Status::state() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.Status.state)
@@ -2737,38 +2855,38 @@ inline std::string* GameServer_Status::mutable_state() {
   return _internal_mutable_state();
 }
 inline const std::string& GameServer_Status::_internal_state() const {
-  return state_.GetNoArena();
+  return state_.Get();
 }
 inline void GameServer_Status::_internal_set_state(const std::string& value) {
   
-  state_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  state_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_Status::set_state(std::string&& value) {
   
-  state_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  state_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.Status.state)
 }
 inline void GameServer_Status::set_state(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  state_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  state_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.Status.state)
 }
-inline void GameServer_Status::set_state(const char* value, size_t size) {
+inline void GameServer_Status::set_state(const char* value,
+    size_t size) {
   
-  state_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  state_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.Status.state)
 }
 inline std::string* GameServer_Status::_internal_mutable_state() {
   
-  return state_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return state_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_Status::release_state() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.Status.state)
-  
-  return state_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return state_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_Status::set_allocated_state(std::string* state) {
   if (state != nullptr) {
@@ -2776,13 +2894,14 @@ inline void GameServer_Status::set_allocated_state(std::string* state) {
   } else {
     
   }
-  state_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), state);
+  state_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), state,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.Status.state)
 }
 
 // string address = 2;
 inline void GameServer_Status::clear_address() {
-  address_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  address_.ClearToEmpty();
 }
 inline const std::string& GameServer_Status::address() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.Status.address)
@@ -2797,38 +2916,38 @@ inline std::string* GameServer_Status::mutable_address() {
   return _internal_mutable_address();
 }
 inline const std::string& GameServer_Status::_internal_address() const {
-  return address_.GetNoArena();
+  return address_.Get();
 }
 inline void GameServer_Status::_internal_set_address(const std::string& value) {
   
-  address_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value);
+  address_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void GameServer_Status::set_address(std::string&& value) {
   
-  address_.SetNoArena(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value));
+  address_.Set(
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:agones.dev.sdk.GameServer.Status.address)
 }
 inline void GameServer_Status::set_address(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  address_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value));
+  address_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:agones.dev.sdk.GameServer.Status.address)
 }
-inline void GameServer_Status::set_address(const char* value, size_t size) {
+inline void GameServer_Status::set_address(const char* value,
+    size_t size) {
   
-  address_.SetNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(reinterpret_cast<const char*>(value), size));
+  address_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:agones.dev.sdk.GameServer.Status.address)
 }
 inline std::string* GameServer_Status::_internal_mutable_address() {
   
-  return address_.MutableNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return address_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* GameServer_Status::release_address() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.Status.address)
-  
-  return address_.ReleaseNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
+  return address_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void GameServer_Status::set_allocated_address(std::string* address) {
   if (address != nullptr) {
@@ -2836,7 +2955,8 @@ inline void GameServer_Status::set_allocated_address(std::string* address) {
   } else {
     
   }
-  address_.SetAllocatedNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), address);
+  address_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), address,
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:agones.dev.sdk.GameServer.Status.address)
 }
 
@@ -2887,21 +3007,43 @@ inline bool GameServer_Status::has_players() const {
   return _internal_has_players();
 }
 inline void GameServer_Status::clear_players() {
-  if (GetArenaNoVirtual() == nullptr && players_ != nullptr) {
+  if (GetArena() == nullptr && players_ != nullptr) {
     delete players_;
   }
   players_ = nullptr;
 }
 inline const ::agones::dev::sdk::GameServer_Status_PlayerStatus& GameServer_Status::_internal_players() const {
   const ::agones::dev::sdk::GameServer_Status_PlayerStatus* p = players_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::agones::dev::sdk::GameServer_Status_PlayerStatus*>(
-      &::agones::dev::sdk::_GameServer_Status_PlayerStatus_default_instance_);
+  return p != nullptr ? *p : reinterpret_cast<const ::agones::dev::sdk::GameServer_Status_PlayerStatus&>(
+      ::agones::dev::sdk::_GameServer_Status_PlayerStatus_default_instance_);
 }
 inline const ::agones::dev::sdk::GameServer_Status_PlayerStatus& GameServer_Status::players() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.Status.players)
   return _internal_players();
 }
+inline void GameServer_Status::unsafe_arena_set_allocated_players(
+    ::agones::dev::sdk::GameServer_Status_PlayerStatus* players) {
+  if (GetArena() == nullptr) {
+    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(players_);
+  }
+  players_ = players;
+  if (players) {
+    
+  } else {
+    
+  }
+  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:agones.dev.sdk.GameServer.Status.players)
+}
 inline ::agones::dev::sdk::GameServer_Status_PlayerStatus* GameServer_Status::release_players() {
+  
+  ::agones::dev::sdk::GameServer_Status_PlayerStatus* temp = players_;
+  players_ = nullptr;
+  if (GetArena() != nullptr) {
+    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  }
+  return temp;
+}
+inline ::agones::dev::sdk::GameServer_Status_PlayerStatus* GameServer_Status::unsafe_arena_release_players() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.Status.players)
   
   ::agones::dev::sdk::GameServer_Status_PlayerStatus* temp = players_;
@@ -2911,7 +3053,7 @@ inline ::agones::dev::sdk::GameServer_Status_PlayerStatus* GameServer_Status::re
 inline ::agones::dev::sdk::GameServer_Status_PlayerStatus* GameServer_Status::_internal_mutable_players() {
   
   if (players_ == nullptr) {
-    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Status_PlayerStatus>(GetArenaNoVirtual());
+    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Status_PlayerStatus>(GetArena());
     players_ = p;
   }
   return players_;
@@ -2921,12 +3063,13 @@ inline ::agones::dev::sdk::GameServer_Status_PlayerStatus* GameServer_Status::mu
   return _internal_mutable_players();
 }
 inline void GameServer_Status::set_allocated_players(::agones::dev::sdk::GameServer_Status_PlayerStatus* players) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArenaNoVirtual();
+  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
   if (message_arena == nullptr) {
     delete players_;
   }
   if (players) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena = nullptr;
+    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
+      ::PROTOBUF_NAMESPACE_ID::Arena::GetArena(players);
     if (message_arena != submessage_arena) {
       players = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
           message_arena, players, submessage_arena);
@@ -2951,21 +3094,43 @@ inline bool GameServer::has_object_meta() const {
   return _internal_has_object_meta();
 }
 inline void GameServer::clear_object_meta() {
-  if (GetArenaNoVirtual() == nullptr && object_meta_ != nullptr) {
+  if (GetArena() == nullptr && object_meta_ != nullptr) {
     delete object_meta_;
   }
   object_meta_ = nullptr;
 }
 inline const ::agones::dev::sdk::GameServer_ObjectMeta& GameServer::_internal_object_meta() const {
   const ::agones::dev::sdk::GameServer_ObjectMeta* p = object_meta_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::agones::dev::sdk::GameServer_ObjectMeta*>(
-      &::agones::dev::sdk::_GameServer_ObjectMeta_default_instance_);
+  return p != nullptr ? *p : reinterpret_cast<const ::agones::dev::sdk::GameServer_ObjectMeta&>(
+      ::agones::dev::sdk::_GameServer_ObjectMeta_default_instance_);
 }
 inline const ::agones::dev::sdk::GameServer_ObjectMeta& GameServer::object_meta() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.object_meta)
   return _internal_object_meta();
 }
+inline void GameServer::unsafe_arena_set_allocated_object_meta(
+    ::agones::dev::sdk::GameServer_ObjectMeta* object_meta) {
+  if (GetArena() == nullptr) {
+    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(object_meta_);
+  }
+  object_meta_ = object_meta;
+  if (object_meta) {
+    
+  } else {
+    
+  }
+  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:agones.dev.sdk.GameServer.object_meta)
+}
 inline ::agones::dev::sdk::GameServer_ObjectMeta* GameServer::release_object_meta() {
+  
+  ::agones::dev::sdk::GameServer_ObjectMeta* temp = object_meta_;
+  object_meta_ = nullptr;
+  if (GetArena() != nullptr) {
+    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  }
+  return temp;
+}
+inline ::agones::dev::sdk::GameServer_ObjectMeta* GameServer::unsafe_arena_release_object_meta() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.object_meta)
   
   ::agones::dev::sdk::GameServer_ObjectMeta* temp = object_meta_;
@@ -2975,7 +3140,7 @@ inline ::agones::dev::sdk::GameServer_ObjectMeta* GameServer::release_object_met
 inline ::agones::dev::sdk::GameServer_ObjectMeta* GameServer::_internal_mutable_object_meta() {
   
   if (object_meta_ == nullptr) {
-    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_ObjectMeta>(GetArenaNoVirtual());
+    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_ObjectMeta>(GetArena());
     object_meta_ = p;
   }
   return object_meta_;
@@ -2985,12 +3150,13 @@ inline ::agones::dev::sdk::GameServer_ObjectMeta* GameServer::mutable_object_met
   return _internal_mutable_object_meta();
 }
 inline void GameServer::set_allocated_object_meta(::agones::dev::sdk::GameServer_ObjectMeta* object_meta) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArenaNoVirtual();
+  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
   if (message_arena == nullptr) {
     delete object_meta_;
   }
   if (object_meta) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena = nullptr;
+    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
+      ::PROTOBUF_NAMESPACE_ID::Arena::GetArena(object_meta);
     if (message_arena != submessage_arena) {
       object_meta = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
           message_arena, object_meta, submessage_arena);
@@ -3011,21 +3177,43 @@ inline bool GameServer::has_spec() const {
   return _internal_has_spec();
 }
 inline void GameServer::clear_spec() {
-  if (GetArenaNoVirtual() == nullptr && spec_ != nullptr) {
+  if (GetArena() == nullptr && spec_ != nullptr) {
     delete spec_;
   }
   spec_ = nullptr;
 }
 inline const ::agones::dev::sdk::GameServer_Spec& GameServer::_internal_spec() const {
   const ::agones::dev::sdk::GameServer_Spec* p = spec_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::agones::dev::sdk::GameServer_Spec*>(
-      &::agones::dev::sdk::_GameServer_Spec_default_instance_);
+  return p != nullptr ? *p : reinterpret_cast<const ::agones::dev::sdk::GameServer_Spec&>(
+      ::agones::dev::sdk::_GameServer_Spec_default_instance_);
 }
 inline const ::agones::dev::sdk::GameServer_Spec& GameServer::spec() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.spec)
   return _internal_spec();
 }
+inline void GameServer::unsafe_arena_set_allocated_spec(
+    ::agones::dev::sdk::GameServer_Spec* spec) {
+  if (GetArena() == nullptr) {
+    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(spec_);
+  }
+  spec_ = spec;
+  if (spec) {
+    
+  } else {
+    
+  }
+  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:agones.dev.sdk.GameServer.spec)
+}
 inline ::agones::dev::sdk::GameServer_Spec* GameServer::release_spec() {
+  
+  ::agones::dev::sdk::GameServer_Spec* temp = spec_;
+  spec_ = nullptr;
+  if (GetArena() != nullptr) {
+    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  }
+  return temp;
+}
+inline ::agones::dev::sdk::GameServer_Spec* GameServer::unsafe_arena_release_spec() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.spec)
   
   ::agones::dev::sdk::GameServer_Spec* temp = spec_;
@@ -3035,7 +3223,7 @@ inline ::agones::dev::sdk::GameServer_Spec* GameServer::release_spec() {
 inline ::agones::dev::sdk::GameServer_Spec* GameServer::_internal_mutable_spec() {
   
   if (spec_ == nullptr) {
-    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Spec>(GetArenaNoVirtual());
+    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Spec>(GetArena());
     spec_ = p;
   }
   return spec_;
@@ -3045,12 +3233,13 @@ inline ::agones::dev::sdk::GameServer_Spec* GameServer::mutable_spec() {
   return _internal_mutable_spec();
 }
 inline void GameServer::set_allocated_spec(::agones::dev::sdk::GameServer_Spec* spec) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArenaNoVirtual();
+  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
   if (message_arena == nullptr) {
     delete spec_;
   }
   if (spec) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena = nullptr;
+    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
+      ::PROTOBUF_NAMESPACE_ID::Arena::GetArena(spec);
     if (message_arena != submessage_arena) {
       spec = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
           message_arena, spec, submessage_arena);
@@ -3071,21 +3260,43 @@ inline bool GameServer::has_status() const {
   return _internal_has_status();
 }
 inline void GameServer::clear_status() {
-  if (GetArenaNoVirtual() == nullptr && status_ != nullptr) {
+  if (GetArena() == nullptr && status_ != nullptr) {
     delete status_;
   }
   status_ = nullptr;
 }
 inline const ::agones::dev::sdk::GameServer_Status& GameServer::_internal_status() const {
   const ::agones::dev::sdk::GameServer_Status* p = status_;
-  return p != nullptr ? *p : *reinterpret_cast<const ::agones::dev::sdk::GameServer_Status*>(
-      &::agones::dev::sdk::_GameServer_Status_default_instance_);
+  return p != nullptr ? *p : reinterpret_cast<const ::agones::dev::sdk::GameServer_Status&>(
+      ::agones::dev::sdk::_GameServer_Status_default_instance_);
 }
 inline const ::agones::dev::sdk::GameServer_Status& GameServer::status() const {
   // @@protoc_insertion_point(field_get:agones.dev.sdk.GameServer.status)
   return _internal_status();
 }
+inline void GameServer::unsafe_arena_set_allocated_status(
+    ::agones::dev::sdk::GameServer_Status* status) {
+  if (GetArena() == nullptr) {
+    delete reinterpret_cast<::PROTOBUF_NAMESPACE_ID::MessageLite*>(status_);
+  }
+  status_ = status;
+  if (status) {
+    
+  } else {
+    
+  }
+  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:agones.dev.sdk.GameServer.status)
+}
 inline ::agones::dev::sdk::GameServer_Status* GameServer::release_status() {
+  
+  ::agones::dev::sdk::GameServer_Status* temp = status_;
+  status_ = nullptr;
+  if (GetArena() != nullptr) {
+    temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
+  }
+  return temp;
+}
+inline ::agones::dev::sdk::GameServer_Status* GameServer::unsafe_arena_release_status() {
   // @@protoc_insertion_point(field_release:agones.dev.sdk.GameServer.status)
   
   ::agones::dev::sdk::GameServer_Status* temp = status_;
@@ -3095,7 +3306,7 @@ inline ::agones::dev::sdk::GameServer_Status* GameServer::release_status() {
 inline ::agones::dev::sdk::GameServer_Status* GameServer::_internal_mutable_status() {
   
   if (status_ == nullptr) {
-    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Status>(GetArenaNoVirtual());
+    auto* p = CreateMaybeMessage<::agones::dev::sdk::GameServer_Status>(GetArena());
     status_ = p;
   }
   return status_;
@@ -3105,12 +3316,13 @@ inline ::agones::dev::sdk::GameServer_Status* GameServer::mutable_status() {
   return _internal_mutable_status();
 }
 inline void GameServer::set_allocated_status(::agones::dev::sdk::GameServer_Status* status) {
-  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArenaNoVirtual();
+  ::PROTOBUF_NAMESPACE_ID::Arena* message_arena = GetArena();
   if (message_arena == nullptr) {
     delete status_;
   }
   if (status) {
-    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena = nullptr;
+    ::PROTOBUF_NAMESPACE_ID::Arena* submessage_arena =
+      ::PROTOBUF_NAMESPACE_ID::Arena::GetArena(status);
     if (message_arena != submessage_arena) {
       status = ::PROTOBUF_NAMESPACE_ID::internal::GetOwnedMessage(
           message_arena, status, submessage_arena);
