@@ -1,4 +1,4 @@
-// Copyright 2020 Google LLC All Rights Reserved.
+// Copyright 2022 Google LLC All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,12 +23,12 @@
 #include <string>
 
 #include <google/protobuf/port_def.inc>
-#if PROTOBUF_VERSION < 3011000
+#if PROTOBUF_VERSION < 3014000
 #error This file was generated by a newer version of protoc which is
 #error incompatible with your Protocol Buffer headers. Please update
 #error your headers.
 #endif
-#if 3011002 < PROTOBUF_MIN_PROTOC_VERSION
+#if 3014000 < PROTOBUF_MIN_PROTOC_VERSION
 #error This file was generated by an older version of protoc which is
 #error incompatible with your Protocol Buffer headers. Please
 #error regenerate this file with a newer version of protoc.
@@ -40,8 +40,7 @@
 #include <google/protobuf/arenastring.h>
 #include <google/protobuf/generated_message_table_driven.h>
 #include <google/protobuf/generated_message_util.h>
-#include <google/protobuf/inlined_string_field.h>
-#include <google/protobuf/metadata.h>
+#include <google/protobuf/metadata_lite.h>
 #include <google/protobuf/generated_message_reflection.h>
 #include <google/protobuf/message.h>
 #include <google/protobuf/repeated_field.h>  // IWYU pragma: export
@@ -60,7 +59,7 @@ PROTOBUF_NAMESPACE_CLOSE
 struct AGONES_EXPORT TableStruct_google_2fapi_2fhttp_2eproto {
   static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTableField entries[]
     PROTOBUF_SECTION_VARIABLE(protodesc_cold);
-  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxillaryParseTableField aux[]
+  static const ::PROTOBUF_NAMESPACE_ID::internal::AuxiliaryParseTableField aux[]
     PROTOBUF_SECTION_VARIABLE(protodesc_cold);
   static const ::PROTOBUF_NAMESPACE_ID::internal::ParseTable schema[3]
     PROTOBUF_SECTION_VARIABLE(protodesc_cold);
@@ -92,10 +91,10 @@ namespace api {
 
 // ===================================================================
 
-class AGONES_EXPORT Http :
+class AGONES_EXPORT Http PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:google.api.Http) */ {
  public:
-  Http();
+  inline Http() : Http(nullptr) {}
   virtual ~Http();
 
   Http(const Http& from);
@@ -109,7 +108,7 @@ class AGONES_EXPORT Http :
     return *this;
   }
   inline Http& operator=(Http&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -117,12 +116,6 @@ class AGONES_EXPORT Http :
     return *this;
   }
 
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArena() const final {
-    return GetArenaNoVirtual();
-  }
-  inline void* GetMaybeArenaPointer() const final {
-    return MaybeArenaPtr();
-  }
   static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
     return GetDescriptor();
   }
@@ -134,7 +127,6 @@ class AGONES_EXPORT Http :
   }
   static const Http& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const Http* internal_default_instance() {
     return reinterpret_cast<const Http*>(
                &_Http_default_instance_);
@@ -147,7 +139,7 @@ class AGONES_EXPORT Http :
   }
   inline void Swap(Http* other) {
     if (other == this) return;
-    if (GetArenaNoVirtual() == other->GetArenaNoVirtual()) {
+    if (GetArena() == other->GetArena()) {
       InternalSwap(other);
     } else {
       ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
@@ -155,7 +147,7 @@ class AGONES_EXPORT Http :
   }
   void UnsafeArenaSwap(Http* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArenaNoVirtual() == other->GetArenaNoVirtual());
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -195,13 +187,6 @@ class AGONES_EXPORT Http :
   private:
   static void ArenaDtor(void* object);
   inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return _internal_metadata_.arena();
-  }
-  inline void* MaybeArenaPtr() const {
-    return _internal_metadata_.raw_arena_ptr();
-  }
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -252,7 +237,6 @@ class AGONES_EXPORT Http :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
   template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
   typedef void InternalArenaConstructable_;
   typedef void DestructorSkippable_;
@@ -263,10 +247,10 @@ class AGONES_EXPORT Http :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT HttpRule :
+class AGONES_EXPORT HttpRule PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:google.api.HttpRule) */ {
  public:
-  HttpRule();
+  inline HttpRule() : HttpRule(nullptr) {}
   virtual ~HttpRule();
 
   HttpRule(const HttpRule& from);
@@ -280,7 +264,7 @@ class AGONES_EXPORT HttpRule :
     return *this;
   }
   inline HttpRule& operator=(HttpRule&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -288,12 +272,6 @@ class AGONES_EXPORT HttpRule :
     return *this;
   }
 
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArena() const final {
-    return GetArenaNoVirtual();
-  }
-  inline void* GetMaybeArenaPointer() const final {
-    return MaybeArenaPtr();
-  }
   static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
     return GetDescriptor();
   }
@@ -315,7 +293,6 @@ class AGONES_EXPORT HttpRule :
     PATTERN_NOT_SET = 0,
   };
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const HttpRule* internal_default_instance() {
     return reinterpret_cast<const HttpRule*>(
                &_HttpRule_default_instance_);
@@ -328,7 +305,7 @@ class AGONES_EXPORT HttpRule :
   }
   inline void Swap(HttpRule* other) {
     if (other == this) return;
-    if (GetArenaNoVirtual() == other->GetArenaNoVirtual()) {
+    if (GetArena() == other->GetArena()) {
       InternalSwap(other);
     } else {
       ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
@@ -336,7 +313,7 @@ class AGONES_EXPORT HttpRule :
   }
   void UnsafeArenaSwap(HttpRule* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArenaNoVirtual() == other->GetArenaNoVirtual());
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -376,13 +353,6 @@ class AGONES_EXPORT HttpRule :
   private:
   static void ArenaDtor(void* object);
   inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return _internal_metadata_.arena();
-  }
-  inline void* MaybeArenaPtr() const {
-    return _internal_metadata_.raw_arena_ptr();
-  }
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -438,15 +408,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_selector();
   std::string* release_selector();
   void set_allocated_selector(std::string* selector);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_selector();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_selector(
-      std::string* selector);
   private:
   const std::string& _internal_selector() const;
   void _internal_set_selector(const std::string& value);
@@ -463,15 +424,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_body();
   std::string* release_body();
   void set_allocated_body(std::string* body);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_body();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_body(
-      std::string* body);
   private:
   const std::string& _internal_body() const;
   void _internal_set_body(const std::string& value);
@@ -488,15 +440,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_response_body();
   std::string* release_response_body();
   void set_allocated_response_body(std::string* response_body);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_response_body();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_response_body(
-      std::string* response_body);
   private:
   const std::string& _internal_response_body() const;
   void _internal_set_response_body(const std::string& value);
@@ -516,15 +459,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_get();
   std::string* release_get();
   void set_allocated_get(std::string* get);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_get();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_get(
-      std::string* get);
   private:
   const std::string& _internal_get() const;
   void _internal_set_get(const std::string& value);
@@ -544,15 +478,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_put();
   std::string* release_put();
   void set_allocated_put(std::string* put);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_put();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_put(
-      std::string* put);
   private:
   const std::string& _internal_put() const;
   void _internal_set_put(const std::string& value);
@@ -572,15 +497,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_post();
   std::string* release_post();
   void set_allocated_post(std::string* post);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_post();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_post(
-      std::string* post);
   private:
   const std::string& _internal_post() const;
   void _internal_set_post(const std::string& value);
@@ -600,15 +516,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_delete_();
   std::string* release_delete_();
   void set_allocated_delete_(std::string* delete_);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_delete_();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_delete_(
-      std::string* delete_);
   private:
   const std::string& _internal_delete_() const;
   void _internal_set_delete_(const std::string& value);
@@ -628,15 +535,6 @@ class AGONES_EXPORT HttpRule :
   std::string* mutable_patch();
   std::string* release_patch();
   void set_allocated_patch(std::string* patch);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_patch();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_patch(
-      std::string* patch);
   private:
   const std::string& _internal_patch() const;
   void _internal_set_patch(const std::string& value);
@@ -676,7 +574,6 @@ class AGONES_EXPORT HttpRule :
   inline bool has_pattern() const;
   inline void clear_has_pattern();
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
   template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
   typedef void InternalArenaConstructable_;
   typedef void DestructorSkippable_;
@@ -700,10 +597,10 @@ class AGONES_EXPORT HttpRule :
 };
 // -------------------------------------------------------------------
 
-class AGONES_EXPORT CustomHttpPattern :
+class AGONES_EXPORT CustomHttpPattern PROTOBUF_FINAL :
     public ::PROTOBUF_NAMESPACE_ID::Message /* @@protoc_insertion_point(class_definition:google.api.CustomHttpPattern) */ {
  public:
-  CustomHttpPattern();
+  inline CustomHttpPattern() : CustomHttpPattern(nullptr) {}
   virtual ~CustomHttpPattern();
 
   CustomHttpPattern(const CustomHttpPattern& from);
@@ -717,7 +614,7 @@ class AGONES_EXPORT CustomHttpPattern :
     return *this;
   }
   inline CustomHttpPattern& operator=(CustomHttpPattern&& from) noexcept {
-    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+    if (GetArena() == from.GetArena()) {
       if (this != &from) InternalSwap(&from);
     } else {
       CopyFrom(from);
@@ -725,12 +622,6 @@ class AGONES_EXPORT CustomHttpPattern :
     return *this;
   }
 
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArena() const final {
-    return GetArenaNoVirtual();
-  }
-  inline void* GetMaybeArenaPointer() const final {
-    return MaybeArenaPtr();
-  }
   static const ::PROTOBUF_NAMESPACE_ID::Descriptor* descriptor() {
     return GetDescriptor();
   }
@@ -742,7 +633,6 @@ class AGONES_EXPORT CustomHttpPattern :
   }
   static const CustomHttpPattern& default_instance();
 
-  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
   static inline const CustomHttpPattern* internal_default_instance() {
     return reinterpret_cast<const CustomHttpPattern*>(
                &_CustomHttpPattern_default_instance_);
@@ -755,7 +645,7 @@ class AGONES_EXPORT CustomHttpPattern :
   }
   inline void Swap(CustomHttpPattern* other) {
     if (other == this) return;
-    if (GetArenaNoVirtual() == other->GetArenaNoVirtual()) {
+    if (GetArena() == other->GetArena()) {
       InternalSwap(other);
     } else {
       ::PROTOBUF_NAMESPACE_ID::internal::GenericSwap(this, other);
@@ -763,7 +653,7 @@ class AGONES_EXPORT CustomHttpPattern :
   }
   void UnsafeArenaSwap(CustomHttpPattern* other) {
     if (other == this) return;
-    GOOGLE_DCHECK(GetArenaNoVirtual() == other->GetArenaNoVirtual());
+    GOOGLE_DCHECK(GetArena() == other->GetArena());
     InternalSwap(other);
   }
 
@@ -803,13 +693,6 @@ class AGONES_EXPORT CustomHttpPattern :
   private:
   static void ArenaDtor(void* object);
   inline void RegisterArenaDtor(::PROTOBUF_NAMESPACE_ID::Arena* arena);
-  private:
-  inline ::PROTOBUF_NAMESPACE_ID::Arena* GetArenaNoVirtual() const {
-    return _internal_metadata_.arena();
-  }
-  inline void* MaybeArenaPtr() const {
-    return _internal_metadata_.raw_arena_ptr();
-  }
   public:
 
   ::PROTOBUF_NAMESPACE_ID::Metadata GetMetadata() const final;
@@ -839,15 +722,6 @@ class AGONES_EXPORT CustomHttpPattern :
   std::string* mutable_kind();
   std::string* release_kind();
   void set_allocated_kind(std::string* kind);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_kind();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_kind(
-      std::string* kind);
   private:
   const std::string& _internal_kind() const;
   void _internal_set_kind(const std::string& value);
@@ -864,15 +738,6 @@ class AGONES_EXPORT CustomHttpPattern :
   std::string* mutable_path();
   std::string* release_path();
   void set_allocated_path(std::string* path);
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  std::string* unsafe_arena_release_path();
-  GOOGLE_PROTOBUF_RUNTIME_DEPRECATED("The unsafe_arena_ accessors for"
-  "    string fields are deprecated and will be removed in a"
-  "    future release.")
-  void unsafe_arena_set_allocated_path(
-      std::string* path);
   private:
   const std::string& _internal_path() const;
   void _internal_set_path(const std::string& value);
@@ -883,7 +748,6 @@ class AGONES_EXPORT CustomHttpPattern :
  private:
   class _Internal;
 
-  ::PROTOBUF_NAMESPACE_ID::internal::InternalMetadataWithArena _internal_metadata_;
   template <typename T> friend class ::PROTOBUF_NAMESPACE_ID::Arena::InternalHelper;
   typedef void InternalArenaConstructable_;
   typedef void DestructorSkippable_;
@@ -968,7 +832,7 @@ inline void Http::set_fully_decode_reserved_expansion(bool value) {
 
 // string selector = 1;
 inline void HttpRule::clear_selector() {
-  selector_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  selector_.ClearToEmpty();
 }
 inline const std::string& HttpRule::selector() const {
   // @@protoc_insertion_point(field_get:google.api.HttpRule.selector)
@@ -987,36 +851,34 @@ inline const std::string& HttpRule::_internal_selector() const {
 }
 inline void HttpRule::_internal_set_selector(const std::string& value) {
   
-  selector_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArenaNoVirtual());
+  selector_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_selector(std::string&& value) {
   
   selector_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.selector)
 }
 inline void HttpRule::set_selector(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  selector_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArenaNoVirtual());
+  selector_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.selector)
 }
 inline void HttpRule::set_selector(const char* value,
     size_t size) {
   
-  selector_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArenaNoVirtual());
+  selector_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.selector)
 }
 inline std::string* HttpRule::_internal_mutable_selector() {
   
-  return selector_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return selector_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_selector() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.selector)
-  
-  return selector_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return selector_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void HttpRule::set_allocated_selector(std::string* selector) {
   if (selector != nullptr) {
@@ -1025,27 +887,8 @@ inline void HttpRule::set_allocated_selector(std::string* selector) {
     
   }
   selector_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), selector,
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.selector)
-}
-inline std::string* HttpRule::unsafe_arena_release_selector() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.selector)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  
-  return selector_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
-}
-inline void HttpRule::unsafe_arena_set_allocated_selector(
-    std::string* selector) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (selector != nullptr) {
-    
-  } else {
-    
-  }
-  selector_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      selector, GetArenaNoVirtual());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.selector)
 }
 
 // string get = 2;
@@ -1057,8 +900,7 @@ inline void HttpRule::set_has_get() {
 }
 inline void HttpRule::clear_get() {
   if (_internal_has_get()) {
-    pattern_.get_.Destroy(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    pattern_.get_.Destroy(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
     clear_has_pattern();
   }
 }
@@ -1078,7 +920,7 @@ inline const std::string& HttpRule::_internal_get() const {
   if (_internal_has_get()) {
     return pattern_.get_.Get();
   }
-  return *&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
+  return ::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
 }
 inline void HttpRule::_internal_set_get(const std::string& value) {
   if (!_internal_has_get()) {
@@ -1086,8 +928,7 @@ inline void HttpRule::_internal_set_get(const std::string& value) {
     set_has_get();
     pattern_.get_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.get_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value,
-      GetArenaNoVirtual());
+  pattern_.get_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_get(std::string&& value) {
   // @@protoc_insertion_point(field_set:google.api.HttpRule.get)
@@ -1097,7 +938,7 @@ inline void HttpRule::set_get(std::string&& value) {
     pattern_.get_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.get_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.get)
 }
 inline void HttpRule::set_get(const char* value) {
@@ -1107,8 +948,8 @@ inline void HttpRule::set_get(const char* value) {
     set_has_get();
     pattern_.get_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.get_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(value), GetArenaNoVirtual());
+  pattern_.get_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{},
+      ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.get)
 }
 inline void HttpRule::set_get(const char* value,
@@ -1119,9 +960,9 @@ inline void HttpRule::set_get(const char* value,
     pattern_.get_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.get_.Set(
-      &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
       reinterpret_cast<const char*>(value), size),
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.get)
 }
 inline std::string* HttpRule::_internal_mutable_get() {
@@ -1130,15 +971,14 @@ inline std::string* HttpRule::_internal_mutable_get() {
     set_has_get();
     pattern_.get_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  return pattern_.get_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
+  return pattern_.get_.Mutable(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_get() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.get)
   if (_internal_has_get()) {
     clear_has_pattern();
-    return pattern_.get_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    return pattern_.get_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
   } else {
     return nullptr;
   }
@@ -1150,31 +990,12 @@ inline void HttpRule::set_allocated_get(std::string* get) {
   if (get != nullptr) {
     set_has_get();
     pattern_.get_.UnsafeSetDefault(get);
+    ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena();
+    if (arena != nullptr) {
+      arena->Own(get);
+    }
   }
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.get)
-}
-inline std::string* HttpRule::unsafe_arena_release_get() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.get)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (_internal_has_get()) {
-    clear_has_pattern();
-    return pattern_.get_.UnsafeArenaRelease(
-        &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
-  } else {
-    return nullptr;
-  }
-}
-inline void HttpRule::unsafe_arena_set_allocated_get(std::string* get) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (!_internal_has_get()) {
-    pattern_.get_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  }
-  clear_pattern();
-  if (get) {
-    set_has_get();
-    pattern_.get_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), get, GetArenaNoVirtual());
-  }
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.get)
 }
 
 // string put = 3;
@@ -1186,8 +1007,7 @@ inline void HttpRule::set_has_put() {
 }
 inline void HttpRule::clear_put() {
   if (_internal_has_put()) {
-    pattern_.put_.Destroy(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    pattern_.put_.Destroy(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
     clear_has_pattern();
   }
 }
@@ -1207,7 +1027,7 @@ inline const std::string& HttpRule::_internal_put() const {
   if (_internal_has_put()) {
     return pattern_.put_.Get();
   }
-  return *&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
+  return ::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
 }
 inline void HttpRule::_internal_set_put(const std::string& value) {
   if (!_internal_has_put()) {
@@ -1215,8 +1035,7 @@ inline void HttpRule::_internal_set_put(const std::string& value) {
     set_has_put();
     pattern_.put_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.put_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value,
-      GetArenaNoVirtual());
+  pattern_.put_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_put(std::string&& value) {
   // @@protoc_insertion_point(field_set:google.api.HttpRule.put)
@@ -1226,7 +1045,7 @@ inline void HttpRule::set_put(std::string&& value) {
     pattern_.put_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.put_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.put)
 }
 inline void HttpRule::set_put(const char* value) {
@@ -1236,8 +1055,8 @@ inline void HttpRule::set_put(const char* value) {
     set_has_put();
     pattern_.put_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.put_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(value), GetArenaNoVirtual());
+  pattern_.put_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{},
+      ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.put)
 }
 inline void HttpRule::set_put(const char* value,
@@ -1248,9 +1067,9 @@ inline void HttpRule::set_put(const char* value,
     pattern_.put_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.put_.Set(
-      &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
       reinterpret_cast<const char*>(value), size),
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.put)
 }
 inline std::string* HttpRule::_internal_mutable_put() {
@@ -1259,15 +1078,14 @@ inline std::string* HttpRule::_internal_mutable_put() {
     set_has_put();
     pattern_.put_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  return pattern_.put_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
+  return pattern_.put_.Mutable(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_put() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.put)
   if (_internal_has_put()) {
     clear_has_pattern();
-    return pattern_.put_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    return pattern_.put_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
   } else {
     return nullptr;
   }
@@ -1279,31 +1097,12 @@ inline void HttpRule::set_allocated_put(std::string* put) {
   if (put != nullptr) {
     set_has_put();
     pattern_.put_.UnsafeSetDefault(put);
+    ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena();
+    if (arena != nullptr) {
+      arena->Own(put);
+    }
   }
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.put)
-}
-inline std::string* HttpRule::unsafe_arena_release_put() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.put)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (_internal_has_put()) {
-    clear_has_pattern();
-    return pattern_.put_.UnsafeArenaRelease(
-        &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
-  } else {
-    return nullptr;
-  }
-}
-inline void HttpRule::unsafe_arena_set_allocated_put(std::string* put) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (!_internal_has_put()) {
-    pattern_.put_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  }
-  clear_pattern();
-  if (put) {
-    set_has_put();
-    pattern_.put_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), put, GetArenaNoVirtual());
-  }
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.put)
 }
 
 // string post = 4;
@@ -1315,8 +1114,7 @@ inline void HttpRule::set_has_post() {
 }
 inline void HttpRule::clear_post() {
   if (_internal_has_post()) {
-    pattern_.post_.Destroy(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    pattern_.post_.Destroy(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
     clear_has_pattern();
   }
 }
@@ -1336,7 +1134,7 @@ inline const std::string& HttpRule::_internal_post() const {
   if (_internal_has_post()) {
     return pattern_.post_.Get();
   }
-  return *&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
+  return ::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
 }
 inline void HttpRule::_internal_set_post(const std::string& value) {
   if (!_internal_has_post()) {
@@ -1344,8 +1142,7 @@ inline void HttpRule::_internal_set_post(const std::string& value) {
     set_has_post();
     pattern_.post_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.post_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value,
-      GetArenaNoVirtual());
+  pattern_.post_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_post(std::string&& value) {
   // @@protoc_insertion_point(field_set:google.api.HttpRule.post)
@@ -1355,7 +1152,7 @@ inline void HttpRule::set_post(std::string&& value) {
     pattern_.post_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.post_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.post)
 }
 inline void HttpRule::set_post(const char* value) {
@@ -1365,8 +1162,8 @@ inline void HttpRule::set_post(const char* value) {
     set_has_post();
     pattern_.post_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.post_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(value), GetArenaNoVirtual());
+  pattern_.post_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{},
+      ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.post)
 }
 inline void HttpRule::set_post(const char* value,
@@ -1377,9 +1174,9 @@ inline void HttpRule::set_post(const char* value,
     pattern_.post_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.post_.Set(
-      &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
       reinterpret_cast<const char*>(value), size),
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.post)
 }
 inline std::string* HttpRule::_internal_mutable_post() {
@@ -1388,15 +1185,14 @@ inline std::string* HttpRule::_internal_mutable_post() {
     set_has_post();
     pattern_.post_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  return pattern_.post_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
+  return pattern_.post_.Mutable(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_post() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.post)
   if (_internal_has_post()) {
     clear_has_pattern();
-    return pattern_.post_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    return pattern_.post_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
   } else {
     return nullptr;
   }
@@ -1408,31 +1204,12 @@ inline void HttpRule::set_allocated_post(std::string* post) {
   if (post != nullptr) {
     set_has_post();
     pattern_.post_.UnsafeSetDefault(post);
+    ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena();
+    if (arena != nullptr) {
+      arena->Own(post);
+    }
   }
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.post)
-}
-inline std::string* HttpRule::unsafe_arena_release_post() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.post)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (_internal_has_post()) {
-    clear_has_pattern();
-    return pattern_.post_.UnsafeArenaRelease(
-        &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
-  } else {
-    return nullptr;
-  }
-}
-inline void HttpRule::unsafe_arena_set_allocated_post(std::string* post) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (!_internal_has_post()) {
-    pattern_.post_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  }
-  clear_pattern();
-  if (post) {
-    set_has_post();
-    pattern_.post_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), post, GetArenaNoVirtual());
-  }
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.post)
 }
 
 // string delete = 5;
@@ -1444,8 +1221,7 @@ inline void HttpRule::set_has_delete_() {
 }
 inline void HttpRule::clear_delete_() {
   if (_internal_has_delete_()) {
-    pattern_.delete__.Destroy(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    pattern_.delete__.Destroy(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
     clear_has_pattern();
   }
 }
@@ -1465,7 +1241,7 @@ inline const std::string& HttpRule::_internal_delete_() const {
   if (_internal_has_delete_()) {
     return pattern_.delete__.Get();
   }
-  return *&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
+  return ::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
 }
 inline void HttpRule::_internal_set_delete_(const std::string& value) {
   if (!_internal_has_delete_()) {
@@ -1473,8 +1249,7 @@ inline void HttpRule::_internal_set_delete_(const std::string& value) {
     set_has_delete_();
     pattern_.delete__.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.delete__.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value,
-      GetArenaNoVirtual());
+  pattern_.delete__.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_delete_(std::string&& value) {
   // @@protoc_insertion_point(field_set:google.api.HttpRule.delete)
@@ -1484,7 +1259,7 @@ inline void HttpRule::set_delete_(std::string&& value) {
     pattern_.delete__.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.delete__.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.delete)
 }
 inline void HttpRule::set_delete_(const char* value) {
@@ -1494,8 +1269,8 @@ inline void HttpRule::set_delete_(const char* value) {
     set_has_delete_();
     pattern_.delete__.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.delete__.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(value), GetArenaNoVirtual());
+  pattern_.delete__.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{},
+      ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.delete)
 }
 inline void HttpRule::set_delete_(const char* value,
@@ -1506,9 +1281,9 @@ inline void HttpRule::set_delete_(const char* value,
     pattern_.delete__.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.delete__.Set(
-      &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
       reinterpret_cast<const char*>(value), size),
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.delete)
 }
 inline std::string* HttpRule::_internal_mutable_delete_() {
@@ -1517,15 +1292,14 @@ inline std::string* HttpRule::_internal_mutable_delete_() {
     set_has_delete_();
     pattern_.delete__.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  return pattern_.delete__.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
+  return pattern_.delete__.Mutable(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_delete_() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.delete)
   if (_internal_has_delete_()) {
     clear_has_pattern();
-    return pattern_.delete__.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    return pattern_.delete__.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
   } else {
     return nullptr;
   }
@@ -1537,31 +1311,12 @@ inline void HttpRule::set_allocated_delete_(std::string* delete_) {
   if (delete_ != nullptr) {
     set_has_delete_();
     pattern_.delete__.UnsafeSetDefault(delete_);
+    ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena();
+    if (arena != nullptr) {
+      arena->Own(delete_);
+    }
   }
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.delete)
-}
-inline std::string* HttpRule::unsafe_arena_release_delete_() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.delete)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (_internal_has_delete_()) {
-    clear_has_pattern();
-    return pattern_.delete__.UnsafeArenaRelease(
-        &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
-  } else {
-    return nullptr;
-  }
-}
-inline void HttpRule::unsafe_arena_set_allocated_delete_(std::string* delete_) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (!_internal_has_delete_()) {
-    pattern_.delete__.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  }
-  clear_pattern();
-  if (delete_) {
-    set_has_delete_();
-    pattern_.delete__.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), delete_, GetArenaNoVirtual());
-  }
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.delete)
 }
 
 // string patch = 6;
@@ -1573,8 +1328,7 @@ inline void HttpRule::set_has_patch() {
 }
 inline void HttpRule::clear_patch() {
   if (_internal_has_patch()) {
-    pattern_.patch_.Destroy(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    pattern_.patch_.Destroy(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
     clear_has_pattern();
   }
 }
@@ -1594,7 +1348,7 @@ inline const std::string& HttpRule::_internal_patch() const {
   if (_internal_has_patch()) {
     return pattern_.patch_.Get();
   }
-  return *&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
+  return ::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited();
 }
 inline void HttpRule::_internal_set_patch(const std::string& value) {
   if (!_internal_has_patch()) {
@@ -1602,8 +1356,7 @@ inline void HttpRule::_internal_set_patch(const std::string& value) {
     set_has_patch();
     pattern_.patch_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.patch_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value,
-      GetArenaNoVirtual());
+  pattern_.patch_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_patch(std::string&& value) {
   // @@protoc_insertion_point(field_set:google.api.HttpRule.patch)
@@ -1613,7 +1366,7 @@ inline void HttpRule::set_patch(std::string&& value) {
     pattern_.patch_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.patch_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.patch)
 }
 inline void HttpRule::set_patch(const char* value) {
@@ -1623,8 +1376,8 @@ inline void HttpRule::set_patch(const char* value) {
     set_has_patch();
     pattern_.patch_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  pattern_.patch_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      ::std::string(value), GetArenaNoVirtual());
+  pattern_.patch_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{},
+      ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.patch)
 }
 inline void HttpRule::set_patch(const char* value,
@@ -1635,9 +1388,9 @@ inline void HttpRule::set_patch(const char* value,
     pattern_.patch_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
   pattern_.patch_.Set(
-      &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
       reinterpret_cast<const char*>(value), size),
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.patch)
 }
 inline std::string* HttpRule::_internal_mutable_patch() {
@@ -1646,15 +1399,14 @@ inline std::string* HttpRule::_internal_mutable_patch() {
     set_has_patch();
     pattern_.patch_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   }
-  return pattern_.patch_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
+  return pattern_.patch_.Mutable(
+      ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_patch() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.patch)
   if (_internal_has_patch()) {
     clear_has_pattern();
-    return pattern_.patch_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-        GetArenaNoVirtual());
+    return pattern_.patch_.ReleaseNonDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
   } else {
     return nullptr;
   }
@@ -1666,31 +1418,12 @@ inline void HttpRule::set_allocated_patch(std::string* patch) {
   if (patch != nullptr) {
     set_has_patch();
     pattern_.patch_.UnsafeSetDefault(patch);
+    ::PROTOBUF_NAMESPACE_ID::Arena* arena = GetArena();
+    if (arena != nullptr) {
+      arena->Own(patch);
+    }
   }
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.patch)
-}
-inline std::string* HttpRule::unsafe_arena_release_patch() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.patch)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (_internal_has_patch()) {
-    clear_has_pattern();
-    return pattern_.patch_.UnsafeArenaRelease(
-        &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
-  } else {
-    return nullptr;
-  }
-}
-inline void HttpRule::unsafe_arena_set_allocated_patch(std::string* patch) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (!_internal_has_patch()) {
-    pattern_.patch_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  }
-  clear_pattern();
-  if (patch) {
-    set_has_patch();
-    pattern_.patch_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), patch, GetArenaNoVirtual());
-  }
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.patch)
 }
 
 // .google.api.CustomHttpPattern custom = 8;
@@ -1705,7 +1438,7 @@ inline void HttpRule::set_has_custom() {
 }
 inline void HttpRule::clear_custom() {
   if (_internal_has_custom()) {
-    if (GetArenaNoVirtual() == nullptr) {
+    if (GetArena() == nullptr) {
       delete pattern_.custom_;
     }
     clear_has_pattern();
@@ -1716,7 +1449,7 @@ inline ::google::api::CustomHttpPattern* HttpRule::release_custom() {
   if (_internal_has_custom()) {
     clear_has_pattern();
       ::google::api::CustomHttpPattern* temp = pattern_.custom_;
-    if (GetArenaNoVirtual() != nullptr) {
+    if (GetArena() != nullptr) {
       temp = ::PROTOBUF_NAMESPACE_ID::internal::DuplicateIfNonNull(temp);
     }
     pattern_.custom_ = nullptr;
@@ -1728,7 +1461,7 @@ inline ::google::api::CustomHttpPattern* HttpRule::release_custom() {
 inline const ::google::api::CustomHttpPattern& HttpRule::_internal_custom() const {
   return _internal_has_custom()
       ? *pattern_.custom_
-      : *reinterpret_cast< ::google::api::CustomHttpPattern*>(&::google::api::_CustomHttpPattern_default_instance_);
+      : reinterpret_cast< ::google::api::CustomHttpPattern&>(::google::api::_CustomHttpPattern_default_instance_);
 }
 inline const ::google::api::CustomHttpPattern& HttpRule::custom() const {
   // @@protoc_insertion_point(field_get:google.api.HttpRule.custom)
@@ -1757,8 +1490,7 @@ inline ::google::api::CustomHttpPattern* HttpRule::_internal_mutable_custom() {
   if (!_internal_has_custom()) {
     clear_pattern();
     set_has_custom();
-    pattern_.custom_ = CreateMaybeMessage< ::google::api::CustomHttpPattern >(
-        GetArenaNoVirtual());
+    pattern_.custom_ = CreateMaybeMessage< ::google::api::CustomHttpPattern >(GetArena());
   }
   return pattern_.custom_;
 }
@@ -1769,7 +1501,7 @@ inline ::google::api::CustomHttpPattern* HttpRule::mutable_custom() {
 
 // string body = 7;
 inline void HttpRule::clear_body() {
-  body_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  body_.ClearToEmpty();
 }
 inline const std::string& HttpRule::body() const {
   // @@protoc_insertion_point(field_get:google.api.HttpRule.body)
@@ -1788,36 +1520,34 @@ inline const std::string& HttpRule::_internal_body() const {
 }
 inline void HttpRule::_internal_set_body(const std::string& value) {
   
-  body_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArenaNoVirtual());
+  body_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_body(std::string&& value) {
   
   body_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.body)
 }
 inline void HttpRule::set_body(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  body_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArenaNoVirtual());
+  body_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.body)
 }
 inline void HttpRule::set_body(const char* value,
     size_t size) {
   
-  body_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArenaNoVirtual());
+  body_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.body)
 }
 inline std::string* HttpRule::_internal_mutable_body() {
   
-  return body_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return body_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_body() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.body)
-  
-  return body_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return body_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void HttpRule::set_allocated_body(std::string* body) {
   if (body != nullptr) {
@@ -1826,32 +1556,13 @@ inline void HttpRule::set_allocated_body(std::string* body) {
     
   }
   body_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), body,
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.body)
-}
-inline std::string* HttpRule::unsafe_arena_release_body() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.body)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  
-  return body_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
-}
-inline void HttpRule::unsafe_arena_set_allocated_body(
-    std::string* body) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (body != nullptr) {
-    
-  } else {
-    
-  }
-  body_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      body, GetArenaNoVirtual());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.body)
 }
 
 // string response_body = 12;
 inline void HttpRule::clear_response_body() {
-  response_body_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  response_body_.ClearToEmpty();
 }
 inline const std::string& HttpRule::response_body() const {
   // @@protoc_insertion_point(field_get:google.api.HttpRule.response_body)
@@ -1870,36 +1581,34 @@ inline const std::string& HttpRule::_internal_response_body() const {
 }
 inline void HttpRule::_internal_set_response_body(const std::string& value) {
   
-  response_body_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArenaNoVirtual());
+  response_body_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void HttpRule::set_response_body(std::string&& value) {
   
   response_body_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.HttpRule.response_body)
 }
 inline void HttpRule::set_response_body(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  response_body_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArenaNoVirtual());
+  response_body_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.HttpRule.response_body)
 }
 inline void HttpRule::set_response_body(const char* value,
     size_t size) {
   
-  response_body_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArenaNoVirtual());
+  response_body_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.HttpRule.response_body)
 }
 inline std::string* HttpRule::_internal_mutable_response_body() {
   
-  return response_body_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return response_body_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* HttpRule::release_response_body() {
   // @@protoc_insertion_point(field_release:google.api.HttpRule.response_body)
-  
-  return response_body_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return response_body_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void HttpRule::set_allocated_response_body(std::string* response_body) {
   if (response_body != nullptr) {
@@ -1908,27 +1617,8 @@ inline void HttpRule::set_allocated_response_body(std::string* response_body) {
     
   }
   response_body_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), response_body,
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:google.api.HttpRule.response_body)
-}
-inline std::string* HttpRule::unsafe_arena_release_response_body() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.HttpRule.response_body)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  
-  return response_body_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
-}
-inline void HttpRule::unsafe_arena_set_allocated_response_body(
-    std::string* response_body) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (response_body != nullptr) {
-    
-  } else {
-    
-  }
-  response_body_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      response_body, GetArenaNoVirtual());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.HttpRule.response_body)
 }
 
 // repeated .google.api.HttpRule additional_bindings = 11;
@@ -1985,7 +1675,7 @@ inline HttpRule::PatternCase HttpRule::pattern_case() const {
 
 // string kind = 1;
 inline void CustomHttpPattern::clear_kind() {
-  kind_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  kind_.ClearToEmpty();
 }
 inline const std::string& CustomHttpPattern::kind() const {
   // @@protoc_insertion_point(field_get:google.api.CustomHttpPattern.kind)
@@ -2004,36 +1694,34 @@ inline const std::string& CustomHttpPattern::_internal_kind() const {
 }
 inline void CustomHttpPattern::_internal_set_kind(const std::string& value) {
   
-  kind_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArenaNoVirtual());
+  kind_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void CustomHttpPattern::set_kind(std::string&& value) {
   
   kind_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.CustomHttpPattern.kind)
 }
 inline void CustomHttpPattern::set_kind(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  kind_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArenaNoVirtual());
+  kind_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.CustomHttpPattern.kind)
 }
 inline void CustomHttpPattern::set_kind(const char* value,
     size_t size) {
   
-  kind_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArenaNoVirtual());
+  kind_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.CustomHttpPattern.kind)
 }
 inline std::string* CustomHttpPattern::_internal_mutable_kind() {
   
-  return kind_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return kind_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* CustomHttpPattern::release_kind() {
   // @@protoc_insertion_point(field_release:google.api.CustomHttpPattern.kind)
-  
-  return kind_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return kind_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void CustomHttpPattern::set_allocated_kind(std::string* kind) {
   if (kind != nullptr) {
@@ -2042,32 +1730,13 @@ inline void CustomHttpPattern::set_allocated_kind(std::string* kind) {
     
   }
   kind_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), kind,
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:google.api.CustomHttpPattern.kind)
-}
-inline std::string* CustomHttpPattern::unsafe_arena_release_kind() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.CustomHttpPattern.kind)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  
-  return kind_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
-}
-inline void CustomHttpPattern::unsafe_arena_set_allocated_kind(
-    std::string* kind) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (kind != nullptr) {
-    
-  } else {
-    
-  }
-  kind_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      kind, GetArenaNoVirtual());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.CustomHttpPattern.kind)
 }
 
 // string path = 2;
 inline void CustomHttpPattern::clear_path() {
-  path_.ClearToEmpty(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  path_.ClearToEmpty();
 }
 inline const std::string& CustomHttpPattern::path() const {
   // @@protoc_insertion_point(field_get:google.api.CustomHttpPattern.path)
@@ -2086,36 +1755,34 @@ inline const std::string& CustomHttpPattern::_internal_path() const {
 }
 inline void CustomHttpPattern::_internal_set_path(const std::string& value) {
   
-  path_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), value, GetArenaNoVirtual());
+  path_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, value, GetArena());
 }
 inline void CustomHttpPattern::set_path(std::string&& value) {
   
   path_.Set(
-    &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::move(value), GetArenaNoVirtual());
+    ::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::move(value), GetArena());
   // @@protoc_insertion_point(field_set_rvalue:google.api.CustomHttpPattern.path)
 }
 inline void CustomHttpPattern::set_path(const char* value) {
   GOOGLE_DCHECK(value != nullptr);
   
-  path_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(value),
-              GetArenaNoVirtual());
+  path_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(value), GetArena());
   // @@protoc_insertion_point(field_set_char:google.api.CustomHttpPattern.path)
 }
 inline void CustomHttpPattern::set_path(const char* value,
     size_t size) {
   
-  path_.Set(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), ::std::string(
-      reinterpret_cast<const char*>(value), size), GetArenaNoVirtual());
+  path_.Set(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, ::std::string(
+      reinterpret_cast<const char*>(value), size), GetArena());
   // @@protoc_insertion_point(field_set_pointer:google.api.CustomHttpPattern.path)
 }
 inline std::string* CustomHttpPattern::_internal_mutable_path() {
   
-  return path_.Mutable(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return path_.Mutable(::PROTOBUF_NAMESPACE_ID::internal::ArenaStringPtr::EmptyDefault{}, GetArena());
 }
 inline std::string* CustomHttpPattern::release_path() {
   // @@protoc_insertion_point(field_release:google.api.CustomHttpPattern.path)
-  
-  return path_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArenaNoVirtual());
+  return path_.Release(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), GetArena());
 }
 inline void CustomHttpPattern::set_allocated_path(std::string* path) {
   if (path != nullptr) {
@@ -2124,27 +1791,8 @@ inline void CustomHttpPattern::set_allocated_path(std::string* path) {
     
   }
   path_.SetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(), path,
-      GetArenaNoVirtual());
+      GetArena());
   // @@protoc_insertion_point(field_set_allocated:google.api.CustomHttpPattern.path)
-}
-inline std::string* CustomHttpPattern::unsafe_arena_release_path() {
-  // @@protoc_insertion_point(field_unsafe_arena_release:google.api.CustomHttpPattern.path)
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  
-  return path_.UnsafeArenaRelease(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      GetArenaNoVirtual());
-}
-inline void CustomHttpPattern::unsafe_arena_set_allocated_path(
-    std::string* path) {
-  GOOGLE_DCHECK(GetArenaNoVirtual() != nullptr);
-  if (path != nullptr) {
-    
-  } else {
-    
-  }
-  path_.UnsafeArenaSetAllocated(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
-      path, GetArenaNoVirtual());
-  // @@protoc_insertion_point(field_unsafe_arena_set_allocated:google.api.CustomHttpPattern.path)
 }
 
 #ifdef __GNUC__
