@@ -30,11 +30,41 @@ namespace Agones.Dev.Sdk.Alpha {
   {
     static readonly string __ServiceName = "agones.dev.sdk.alpha.SDK";
 
-    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.PlayerID> __Marshaller_agones_dev_sdk_alpha_PlayerID = grpc::Marshallers.Create((arg) => global::Google.Protobuf.MessageExtensions.ToByteArray(arg), global::Agones.Dev.Sdk.Alpha.PlayerID.Parser.ParseFrom);
-    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.Bool> __Marshaller_agones_dev_sdk_alpha_Bool = grpc::Marshallers.Create((arg) => global::Google.Protobuf.MessageExtensions.ToByteArray(arg), global::Agones.Dev.Sdk.Alpha.Bool.Parser.ParseFrom);
-    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.Count> __Marshaller_agones_dev_sdk_alpha_Count = grpc::Marshallers.Create((arg) => global::Google.Protobuf.MessageExtensions.ToByteArray(arg), global::Agones.Dev.Sdk.Alpha.Count.Parser.ParseFrom);
-    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.Empty> __Marshaller_agones_dev_sdk_alpha_Empty = grpc::Marshallers.Create((arg) => global::Google.Protobuf.MessageExtensions.ToByteArray(arg), global::Agones.Dev.Sdk.Alpha.Empty.Parser.ParseFrom);
-    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.PlayerIDList> __Marshaller_agones_dev_sdk_alpha_PlayerIDList = grpc::Marshallers.Create((arg) => global::Google.Protobuf.MessageExtensions.ToByteArray(arg), global::Agones.Dev.Sdk.Alpha.PlayerIDList.Parser.ParseFrom);
+    static void __Helper_SerializeMessage(global::Google.Protobuf.IMessage message, grpc::SerializationContext context)
+    {
+      #if !GRPC_DISABLE_PROTOBUF_BUFFER_SERIALIZATION
+      if (message is global::Google.Protobuf.IBufferMessage)
+      {
+        context.SetPayloadLength(message.CalculateSize());
+        global::Google.Protobuf.MessageExtensions.WriteTo(message, context.GetBufferWriter());
+        context.Complete();
+        return;
+      }
+      #endif
+      context.Complete(global::Google.Protobuf.MessageExtensions.ToByteArray(message));
+    }
+
+    static class __Helper_MessageCache<T>
+    {
+      public static readonly bool IsBufferMessage = global::System.Reflection.IntrospectionExtensions.GetTypeInfo(typeof(global::Google.Protobuf.IBufferMessage)).IsAssignableFrom(typeof(T));
+    }
+
+    static T __Helper_DeserializeMessage<T>(grpc::DeserializationContext context, global::Google.Protobuf.MessageParser<T> parser) where T : global::Google.Protobuf.IMessage<T>
+    {
+      #if !GRPC_DISABLE_PROTOBUF_BUFFER_SERIALIZATION
+      if (__Helper_MessageCache<T>.IsBufferMessage)
+      {
+        return parser.ParseFrom(context.PayloadAsReadOnlySequence());
+      }
+      #endif
+      return parser.ParseFrom(context.PayloadAsNewBuffer());
+    }
+
+    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.PlayerID> __Marshaller_agones_dev_sdk_alpha_PlayerID = grpc::Marshallers.Create(__Helper_SerializeMessage, context => __Helper_DeserializeMessage(context, global::Agones.Dev.Sdk.Alpha.PlayerID.Parser));
+    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.Bool> __Marshaller_agones_dev_sdk_alpha_Bool = grpc::Marshallers.Create(__Helper_SerializeMessage, context => __Helper_DeserializeMessage(context, global::Agones.Dev.Sdk.Alpha.Bool.Parser));
+    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.Count> __Marshaller_agones_dev_sdk_alpha_Count = grpc::Marshallers.Create(__Helper_SerializeMessage, context => __Helper_DeserializeMessage(context, global::Agones.Dev.Sdk.Alpha.Count.Parser));
+    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.Empty> __Marshaller_agones_dev_sdk_alpha_Empty = grpc::Marshallers.Create(__Helper_SerializeMessage, context => __Helper_DeserializeMessage(context, global::Agones.Dev.Sdk.Alpha.Empty.Parser));
+    static readonly grpc::Marshaller<global::Agones.Dev.Sdk.Alpha.PlayerIDList> __Marshaller_agones_dev_sdk_alpha_PlayerIDList = grpc::Marshallers.Create(__Helper_SerializeMessage, context => __Helper_DeserializeMessage(context, global::Agones.Dev.Sdk.Alpha.PlayerIDList.Parser));
 
     static readonly grpc::Method<global::Agones.Dev.Sdk.Alpha.PlayerID, global::Agones.Dev.Sdk.Alpha.Bool> __Method_PlayerConnect = new grpc::Method<global::Agones.Dev.Sdk.Alpha.PlayerID, global::Agones.Dev.Sdk.Alpha.Bool>(
         grpc::MethodType.Unary,
@@ -92,6 +122,7 @@ namespace Agones.Dev.Sdk.Alpha {
     }
 
     /// <summary>Base class for server-side implementations of SDK</summary>
+    [grpc::BindServiceMethod(typeof(SDK), "BindService")]
     public abstract partial class SDKBase
     {
       /// <summary>
@@ -217,7 +248,7 @@ namespace Agones.Dev.Sdk.Alpha {
     {
       /// <summary>Creates a new client for SDK</summary>
       /// <param name="channel">The channel to use to make remote calls.</param>
-      public SDKClient(grpc::Channel channel) : base(channel)
+      public SDKClient(grpc::ChannelBase channel) : base(channel)
       {
       }
       /// <summary>Creates a new client for SDK that uses a custom <c>CallInvoker</c>.</summary>
